% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-utils.R
\name{add_data_sheet}
\alias{add_data_sheet}
\title{Wrappers for openxlsx pacakge functions}
\usage{
add_data_sheet(
  wb,
  data,
  sheetname,
  tableStyle = "TableStyleLight8",
  bandedRows = TRUE,
  bandedCols = TRUE,
  ...,
  override = TRUE
)
}
\arguments{
\item{wb}{A Workbook object to attach the new worksheet and table}

\item{data}{A dataframe.}

\item{sheetname}{The worksheet to write to. Can be the worksheet index or name.}

\item{tableStyle}{Any excel table style name or "none".}

\item{bandedRows}{Logical. If TRUE, rows are colour banded}

\item{bandedCols}{Logical. If TRUE, the columns are colour banded}

\item{...}{Additional arguments passed to openxlsx::writeDataTable that are not already listed}

\item{override}{Logical.  If TRUE, will delete the sheetname (if present)}
}
\description{
Add data to sheet
}
\details{
A wrapper function to use within the openxlsx workbook building.
This adds additional functionality to override within the function and provude an output that can be piped.

These should be applied to a string starting with [openxlsx::createWorkbook()] then piped through with the functions below.
}
