% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{add_euclidean}
\alias{add_euclidean}
\title{Add Euclidean distance}
\usage{
add_euclidean(df, .x, .y, .name = "eucd")
}
\arguments{
\item{df}{A data.frame}

\item{.x, .y}{The two values to compute the distance from}

\item{.name}{The name of the new column for the distance value}
}
\description{
Add a Euclidean distance measure to the end of a data frame
}
\examples{
add_euclidean(iris, "Sepal.Width", "Sepal.Length")
}
