% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{add_mahalanobis}
\alias{add_mahalanobis}
\title{Add Mahalanobis distance}
\usage{
add_mahalanobis(
  df,
  ...,
  .inverted = FALSE,
  .name = "md",
  .p = "p_value",
  tolerance = .Machine$double.eps
)
}
\arguments{
\item{df}{A data.frame}

\item{...}{Columns to select}

\item{.inverted}{Logical. If TRUE, covariance matrix (p x p) of the distribution is supposed to contain the inverse of the covariance matrix.}

\item{.name}{The name of the new column for the distance value}

\item{.p}{The name for the new column for the p-value calculation}

\item{tolerance}{\code{tol} in \code{\link[base:solve]{base::solve()}}}
}
\description{
Add a mahalanobis distance to the end of a data frame.
}
\examples{
add_mahalanobis(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width) \%>\% head
df <- iris
df[1, ] <- NA_real_
head(df)
add_mahalanobis(df, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width) \%>\% head
}
