% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-base.R
\name{alpha_base}
\alias{alpha_base}
\title{Alpha base}
\usage{
alpha_base(x)
}
\arguments{
\item{x}{A string of letters.  Non characters are removed.}
}
\description{
Base 26 conversion with letters
}
\examples{
alpha_base("AB")
alpha_base("XFD")
alpha_base(c("JMB", "Jordan Mark", "XKCD"))
}
