% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{as.integer.fact}
\alias{as.integer.fact}
\title{#' @export
is.na.fact <- function(x) {
unclass(x) == attr(x, "na")
}}
\usage{
\method{as.integer}{fact}(x, ...)
}
\description{
#' @export
is.na.fact <- function(x) {
unclass(x) == attr(x, "na")
}
}
