% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-add.R
\name{assigns}
\alias{assigns}
\alias{\%=+\%}
\alias{\%=-\%}
\title{Assign add}
\usage{
e1 \%=+\% e2

e1 \%=-\% e2
}
\arguments{
\item{e1}{Object to be updated}

\item{e2}{right side statement}
}
\description{
this needs to be tested...
}
\examples{
val <- 1
val \%=+\% 2
val # 3
val \%=+\% 2
val # 5
val \%=-\% 1
val # 4
\dontrun{
val \%=+\% c(1, 2, 3) # fails
}
val <- c(1, 2, 3)
val \%=+\% c(1, 2, 3) # 2 4 6
}
