% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{attached_packages}
\alias{attached_packages}
\alias{.SendAttachedPackagesToREnviron}
\alias{.RemoveAttachedPackages}
\title{Manage attached packages}
\usage{
.SendAttachedPackagesToREnviron()

.RemoveAttachedPackages(attached = NULL)
}
\arguments{
\item{attached}{A character vector of packages - if NULL will find packages}
}
\description{
Clears out attached packages and properly loads
}
\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{ht}()},
\code{\link{startup_funs}}
}
\concept{startup_utils}
