% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha-base.R
\name{base_alpha}
\alias{base_alpha}
\title{Alpha base}
\usage{
base_alpha(x, base = 26)
}
\arguments{
\item{x}{A string of letters.  Non characters are removed.}

\item{base}{A numeric}
}
\description{
Base 26 conversion with letters
}
\examples{
base_alpha("AB")
base_alpha("XFD")
base_alpha(c("JMB", "Jordan Mark", "XKCD"))
sum(base_alpha(c("x", "k", "c", "d")))
}
