% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-conversion.R
\name{base_n}
\alias{base_n}
\title{Base N conversion}
\usage{
base_n(x, base = 10)
}
\arguments{
\item{x}{A vector of integers}

\item{base}{An integer base (valid 1 through 10)}
}
\description{
Convert between base numbers (1 through 10)
}
