% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-more.R
\name{betwixt}
\alias{betwixt}
\alias{between}
\alias{betwee_more}
\alias{between_more}
\title{Betwixt boundaries}
\usage{
between_more(x, left, right, type = c("gele", "gel", "gle", "gl"))

betwixt(x, left = NULL, right = NULL, bounds = c("[]", "[)", "(]", "()"))
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values.  For \code{\link[=betwixt]{betwixt()}}, when \code{NULL} no
comparison is made for that boundary.  When both are \code{NULL}, \code{x} is just
returned.}

\item{type}{Abbreviation for the evaluation of \code{left} on \code{right} (see
details)}

\item{bounds}{Boundaries for comparisons of \code{left} and \code{right} (see details)}
}
\value{
A logical vector
}
\description{
Compare a vector betwitx (between) other values
}
\details{
\verb{type``, }bounds`` can be one of the below:

\describe{
\item{g,(}{is greater than (>)}
\item{ge,[}{greater than or equal to (>=)}
\item{l,))}{less than (<)}
\item{le,[]}{less than or equal to (<=)}
}

Note: \code{\link[=between_more]{between_more()}} may be deprecated in the future in favor of just
\code{\link[=betwixt]{betwixt()}}
}
\examples{

between_more(2:10, 2, 10, "gl")
betwixt(2:10, 2, bounds = "()")
between_more(10, 2, 10, "gle")
betwixt(2:10, bounds = "(]")
betwixt(1:5, c(3, 3, 2, 2, 1), 5)
}
