% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{bump_date_version}
\alias{bump_date_version}
\title{Bump version (date)}
\usage{
bump_date_version()
}
\description{
Will read the DESCRIPTION file and adjust the version
Assumes that the version is written with major releases of 0 and appending
the date.

When the current version is the same as today's date, it will append a \code{.1}.
}
\details{
Bump package version for dates
}
