% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{checkOptions}
\alias{checkOptions}
\title{Check options}
\usage{
checkOptions(x)
}
\arguments{
\item{x}{A named list of new options}
}
\value{
Invisible, a list of the current options from \code{options()}
}
\description{
For each name in \code{x} checks the current option value and reports if there
is a difference in a \code{message}.  This does not change the options
}
\details{
Checks and reports on options
}
\examples{
op <- options()

x <- list(width = -20, warning.length = 2, probably_not_a_real_option = 2)
checkOptions(x)
# pointless, but shows that no messages are given
identical(options(), checkOptions(options()))

options(op)
}
