% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{chr_split}
\alias{chr_split}
\title{Character split}
\usage{
chr_split(x)
}
\arguments{
\item{x}{A vector of strings to split}
}
\value{
A \code{character} vector of length \code{nchar(x)}
}
\description{
Split apart a string by each character
}
\examples{
chr_split("split this")
}
