% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{cohensd2r}
\alias{cohensd2r}
\title{r to Cohen's D}
\usage{
cohensd2r(d, n1 = NULL, n2 = n1, var = F)
}
\arguments{
\item{d}{Cohen's d}

\item{n1, n2}{Ns for the groups}

\item{var}{Logical.}
}
\description{
Converts an r value to Cohen's D value
}
