% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\alias{counts.data.frame}
\alias{props}
\alias{props.default}
\alias{props.data.frame}
\title{Count observations by unique values}
\usage{
counts(x, ...)

\method{counts}{data.frame}(x, cols, ...)

props(x, ...)

\method{props}{default}(x, ...)

\method{props}{data.frame}(x, cols, ...)
}
\arguments{
\item{x}{A vector or data.frame}

\item{...}{Arguments passed to other methods}

\item{cols}{A vector of column names or indexes}
}
\description{
Variables will be return by the order in which they appear
}
\details{
Get counts or proportions of unique observations in a vector or columns in a
data.frame
}
\examples{
x <- sample(1:5, 10, TRUE)
counts(x)
props(x)

x <- quick_df(list(
  a = c("a", "c", "a", "c", "d", "b"),
  b = c("a", "a", "a", "c", "c", "b"),
  c = c("a", "a", "a", "c", "b", "b")
))

counts(x, "a")
counts(x, c("a", "b", "c"))
props(x, 2)
props(x, 1:3)
}
