% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_from_partial}
\alias{date_from_partial}
\title{Partial dates}
\usage{
date_from_partial(
  x,
  format = "ymd",
  method = c("min", "max"),
  year_replacement = NA
)
}
\arguments{
\item{x}{A vector of character dates}

\item{format}{Format order of the date (accepts only combinations of 'y',
'm', and 'd')}

\item{method}{Method for reporting partial dates -- "min" or "max"}

\item{year_replacement}{If set, will use a year}
}
\description{
Derive a date object from a partial date
}
\details{
Takes a character as an argument and attempts to create a date object when
part of the date string is missing.
}
\examples{
x <- c("2020-12-17", NA_character_, "", "2020-12-UN", "2020-12-UN",
       "2019-Unknown-00", "UNK-UNK-UNK", "1991-02-UN", "    ",
       "2020January20")
data.frame(x = x,
           min = date_from_partial(x),
           max = date_from_partial(x, method = "max"))
}
