% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate_fun}
\alias{deprecate_fun}
\title{Deprecate a function with a message}
\usage{
deprecate_fun(
  old_package,
  old_function,
  new_package = old_package,
  new_function = old_function
)
}
\arguments{
\item{old_package}{Package for old function}

\item{old_function}{Function to deprecate}

\item{new_package}{New package for replacing function}

\item{new_function}{New function to replace}
}
\description{
Adds adds a note that
}
\examples{
deprecate_fun("jordan", "fake_fun", new_function = "real_fun")
# Running a second time won't produce message
deprecate_fun("jordan", "fake_fun", new_function = "real_fun")
}
