% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.R
\name{detail}
\alias{detail}
\alias{detail.default}
\alias{detail.data.frame}
\title{Details an object}
\usage{
detail(x, ...)

\method{detail}{default}(x, ...)

\method{detail}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}
}
\description{
Provides details about an object
}
\examples{
x <- sample(letters[1:4], 10, TRUE)
detail(x)

df <- quick_dfl(
  x = x,
  y = round(runif(10), 2),
  z = Sys.Date() + runif(10) * 100
)

detail(df)
}
