% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{.CharacterIndex}
\alias{.CharacterIndex}
\title{Character index}
\usage{
.CharacterIndex(x = read_clipboard())
}
\arguments{
\item{x}{A character vector}
}
\description{
Splits a vector into each character and shows the index
}
\examples{
x <- c("Split this apart", "and count each character.")
.CharacterIndex(x)

}
\references{
Inspired by: https://www.r-bloggers.com/2020/12/little-helpers-character-index-counter/
}
