% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{drop_levels}
\alias{drop_levels}
\alias{drop_levels.data.frame}
\alias{drop_levels.fact}
\alias{drop_levels.factor}
\title{Drop levels}
\usage{
drop_levels(x, ...)

\method{drop_levels}{data.frame}(x, ...)

\method{drop_levels}{fact}(x, ...)

\method{drop_levels}{factor}(x, ...)
}
\arguments{
\item{x}{A \code{factor} or \code{data.frame}}

\item{...}{Additional arguments passed to methods (not used)}
}
\description{
Drop unused levels of a factor
}
\seealso{
\link[base:droplevels]{base::droplevels}

Other factors:
\code{\link[=as_ordered]{as_ordered()}},
\code{\link[=char2fact]{char2fact()}},
\code{\link[=fact]{fact()}},
\code{\link[=fact2char]{fact2char()}},
\code{\link[=fact_na]{fact_na()}}
}
\concept{factors}
