% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{echo}
\alias{echo}
\title{echo}
\usage{
echo(exprs, to = stdout(), msg = TRUE)
}
\arguments{
\item{exprs}{Expressions}

\item{to}{Output locations}

\item{msg}{If \code{FALSE} does not print results from \code{message()}}
}
\description{
Run expressions with logging outputs
}
\details{
Function is deprecated.  Use \code{echo::echo()} instead.
}
\examples{
if (package_available("echo")) {
  try(echo::echo({
    1 + 1
    Sys.sleep(2)
    print(head(mtcars))
    message(1)
    warning(2)
    stop(3)
  },
  level = 0
  ))
}
}
