% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-rust.R
\name{engine_rust}
\alias{engine_rust}
\alias{set_rust_engine}
\title{Engine - Rust}
\usage{
engine_rust(options)

set_rust_engine()
}
\arguments{
\item{options}{Options passed to code chunk}
}
\description{
Set Rust engine for knitr
}
\details{
If dependencies are needed they should be defined in a `cargo.toml` file in
  the working directory and the code chunk option `toml=TRUE` should be set.
}
\references{
Adapted from
  https://greenwood.dev/2019/12/22/running-rust-in-rmd/
}
