% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-named-chunk.R
\name{eval_named_chunk}
\alias{eval_named_chunk}
\title{Evaluate a  Named Chunk}
\usage{
eval_named_chunk(rmd_file, label_name, ...)
}
\arguments{
\item{rmd_file}{Absolute path to rmd file}

\item{label_name}{Name of label}

\item{...}{Additional arguments passed to [base::eval()]}
}
\description{
Evaluate a named chunk from an Rmd file.
}
\examples{
\dontrun{
temp_rmd <- tempfile(fileext = ".rmd")

text <- '
```{r not this label}
print("that is wrong")
```

```{r hello label}
text <- "hello, world"
print(text)
```

```{r another label}
warning("wrong label")
```
'

writeLines(text, con = temp_rmd)

read_named_chunk(temp_rmd, "hello label")
}
}
