% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand_by}
\alias{expand_by}
\title{Expands a vector}
\usage{
expand_by(x, y, expand = c("x", "y", "intersect", "both"), sort = FALSE)
}
\arguments{
\item{x, y}{Vectors}

\item{expand}{Character switch to expand or keep only the values that
intersect, all values in x or y, or retain all values found.}

\item{sort}{Logical, if \code{TRUE} will sort by names in output}
}
\value{
A vector with expanded
}
\description{
Expands vector x by y
}
\examples{
x <- letters[c(3:2, 5, 9)]
y <- letters[c(1:4, 8)]
expand_by(x, y, "x")
expand_by(x, y, "y")
expand_by(x, y, "intersect")
expand_by(x, y, "both")
}
