% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2fact.R
\name{fact2char}
\alias{fact2char}
\title{Factor to character}
\usage{
fact2char(data, threshold = 10)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{threshold}{A threshold for the number of levels to be met/exceeded for
transforming into a character}
}
\value{
The \code{data.frame} \code{data} with factors converted by the rule above
}
\description{
Convert factor columns to characters in a \code{data.frame}
}
\seealso{
\code{\link[=char2fact]{char2fact()}}

Other factors:
\code{\link[=char2fact]{char2fact()}},
\code{\link[=fact]{fact()}}
}
\concept{factors}
