% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact_na}
\alias{fact_na}
\title{\code{fact} with \code{NA}}
\usage{
fact_na(x, remove = FALSE)
}
\arguments{
\item{x}{A \code{fact} or object cohered to \code{fact}}

\item{remove}{If \code{TRUE} removes \code{NA} value from the \code{fact} \code{levels} and
\code{uniques} attributes}
}
\value{
A \code{fact} vector
}
\description{
Included \code{NA} values into \code{fact()}
}
\details{
This re-formats the \code{x} value so that \code{NA}s are found immediately within the
object rather than accessed through its attributes.
}
\seealso{
Other factors:
\code{\link[=as_ordered]{as_ordered()}},
\code{\link[=char2fact]{char2fact()}},
\code{\link[=drop_levels]{drop_levels()}},
\code{\link[=fact]{fact()}},
\code{\link[=fact2char]{fact2char()}}
}
\concept{factors}
