% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact_reverse}
\alias{fact_reverse}
\title{Fact reverse levels}
\usage{
fact_reverse(x)
}
\arguments{
\item{x}{A \code{fact} object (or passed to \code{\link[=fact]{fact()}})}
}
\description{
Reverse the levels of a \code{fact}
}
