% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_expand_seq.R
\name{fct_expand_seq}
\alias{fct_expand_seq}
\title{Factor Expand by Sequence}
\usage{
fct_expand_seq(
  x,
  min_lvl = min(x, na.rm = TRUE),
  max_lvl = max(x, na.rm = TRUE),
  by = 1L
)
}
\arguments{
\item{x}{An ordered factor}

\item{min_lvl}{The start of the level sequence}

\item{max_lvl}{The end of the level sequence}

\item{by}{Integer, number of steps in between}
}
\description{
Expands an ordered factor from one level to another
}
\details{
Defaults for `min_lvl` and `max_lvl` are the minimum and maximum levels in the ordered vector `x`.
This can be adjusted if we
}
\examples{
x <- factor(letters[c(5:15, 2)], levels = letters, ordered = TRUE)
fct_expand_seq(x)
fct_expand_seq(x, "j", by = 3L)
}
