% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers_method.R
\name{fishers_method}
\alias{fishers_method}
\title{Fisher's method for combined probabilites}
\usage{
fishers_method(.x, ...)
}
\arguments{
\item{.x}{A vector of p-values.}

\item{...}{Additional arguments passed to \link[base][pchisq]}
}
\description{
Computes Fisher's method for combined probabilities
}
