% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\alias{flip.default}
\alias{flip.data.frame}
\title{Flip an object}
\usage{
flip(x, ...)

\method{flip}{default}(x, ...)

\method{flip}{data.frame}(x, by_row = TRUE, rownames = FALSE, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}

\item{by_row}{`TRUE`, flips by row, otherwise by column}

\item{rownames}{`FALSE`, if `TRUE` will update the rownames}
}
\description{
Flip an object
}
\examples{
flip(letters[1:3])
flip(seq.int(9, -9, by = -3))
flip(head(iris))
flip(head(iris), rownames = TRUE)
flip(head(iris), by_row = FALSE)
}
