% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{format_to_regex}
\alias{format_to_regex}
\title{Format string to a regular expression}
\usage{
format_to_regex(x)
}
\arguments{
\item{x}{A date or datetime format, assuming that entries follow the formats
described in \link[base:strptime]{base::strptime}}
}
\description{
Format string to a regular expression
}
\examples{
mark:::format_to_regex("\%Y-\%m-\%d")
mark:::format_to_regex("\%b/\%d/\%y")
mark:::format_to_regex("\%d \%B \%Y")

pattern <- mark:::format_to_regex("\%Y-\%m-\%d")
grepl(pattern, Sys.Date())
}
