% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-report.R
\name{formatTimeDiff}
\alias{formatTimeDiff}
\title{Format time difference}
\usage{
formatTimeDiff(start, stop = Sys.time(), threshold = 0.1)
}
\arguments{
\item{start}{A start time}

\item{stop}{A stop time (default to \code{Sys.time()})}

\item{threshold}{A threshold for reporting the difference, if \code{stop - start}
is less than this, a empty character vector (\code{""}) is returned}
}
\description{
A preferred method of printing out a time differences, assuming it is to be
appended to some other message
}
\details{
Provides quick formatting for time differences object
}
