% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{fp}
\alias{fp}
\title{File path}
\usage{
fp(..., sep = .Platform$file.sep)
}
\arguments{
\item{...}{Character vectors}

\item{sep}{The path separator to be applied}
}
\description{
Create a character string of a file path and check that it exists
}
\details{
File path is created with `file.path()` then passed to [tools::file_path_as_absolute()]
  and then checked with `dir.exists(basename(.))`.
}
