% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir_max_number}
\alias{get_dir_max_number}
\title{Get recent directory by number name}
\usage{
get_dir_max_number(x)
}
\arguments{
\item{x}{The directory to look in}
}
\value{
A full path to a directory
}
\description{
Finds the directory where the number is the greatest.  This can be useful for when folders are created as run IDs.
}
