% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir_recent_date}
\alias{get_dir_recent_date}
\title{Get recent directory by folder name}
\usage{
get_dir_recent_date(dirs, dt_pattern = NULL, dt_format = NULL)
}
\arguments{
\item{dirs}{A directory.}

\item{dt_pattern}{A pattern to be passed to filter for the directory.}

\item{dt_format}{A format to be passed for the date.}
}
\description{
Finds the most recent subdirectory assuming it is a number or date
}
\details{
Get recent directory by date

Looks at the directories and assumes the date
}
