% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{get_recent_dir}
\alias{get_recent_dir}
\title{Get recent directory}
\usage{
get_recent_dir(.dir, recursive = FALSE, ...)
}
\arguments{
\item{.dir}{The root directory}

\item{recursive}{From `list.dirs()`}

\item{...}{Further arguments passed to `list.dirs()`}
}
\value{
The full path of the most recent directory
}
\description{
Finds the recent subdirectory in a directory.
}
