% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R, R/get_recent_file.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Get recent file}
\usage{
get_recent_file(dir, pattern = NULL)

get_recent_file(dir, pattern = NULL)
}
\arguments{
\item{dir}{A directory.}

\item{pattern}{A regular expression to filter for.}
}
\description{
A function where you can detect the most recent file from a directory.

Get the recent file from a directory
}
