% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{get_version}
\alias{get_version}
\alias{bump_version}
\alias{bump_date_version}
\alias{update_version}
\title{Get and bump version}
\usage{
get_version()

bump_version(version = NULL)

bump_date_version(version = NULL)

update_version(version = NULL, date = FALSE)
}
\arguments{
\item{version}{A new version to be added; default of \code{NULL} will
automatically update.}

\item{date}{If \code{TRUE} will use a date as a version.}
}
\value{
\itemize{
\item \code{get_version()}: A package_version
\item \code{bump_version()}: None, called for its side-effects
\item \code{bump_date_version()}: None, called for its side-effects
\item \code{update_version()}: None, called for its side-effects
}
}
\description{
Will read the \code{DESCRIPTION} file and to get and adjust the
version

\code{bump_date_version()} will not check if the version is actually a date.  When
the current version is the same as today's date(equal by character strings)
it will append a \code{.1}.
}
\details{
Get and bump package version for dates
}
