% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob.R
\name{glob}
\alias{glob}
\title{File globbing}
\usage{
glob(x, glob = NULL, regexp = NULL, ...)
}
\arguments{
\item{x}{A vector of characters}

\item{glob, regexp}{Search expressions, as either wildcard globbing or as a
regular expression.  \code{glob} is by default passed to regexp via
\code{\link[utils:glob2rx]{utils::glob2rx()}}}

\item{...}{Additional parameters passed to \code{grep}; \code{value} is ignored}
}
\description{
Helper function for globbing files
}
