% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%colons\%}
\alias{\%colons\%}
\title{Colons}
\usage{
package \%colons\% name
}
\arguments{
\item{package}{Name of the package}

\item{name}{Name to retrieve}
}
\value{
The variable \code{name} from package \code{package}
}
\description{
Get an object from a package
}
\details{
This is a work around to calling \code{:::}.
}
\section{WARNING}{

To reiterate from other documentation: it is not advised to use \code{:::} in
your code as it will retrieve non-exported objects that may be more
likely to change in their functionality that exported objects.
}

