% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-add.R
\name{\%=+\%}
\alias{\%=+\%}
\title{Assign add}
\usage{
lhs \%=+\% rhs
}
\arguments{
\item{lhs}{left side statement}

\item{rhs}{right side statement}
}
\description{
this needs to be tested...
}
\examples{
val <- 1
val \%=+\% 2
val # 3
val \%=+\% 2
val # 5
\dontrun{
val \%=+\% c(1, 2, 3) # fails
}
}
