% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{ht}
\alias{ht}
\title{Head-Tail}
\usage{
ht(x, n = 5L)
}
\arguments{
\item{x}{A data.frame}

\item{n}{The number of rows to see (if length 1L, is repeated for head and
tail)}
}
\description{
Quick head & tail function
}
\details{
This contains a separate method for class \code{tbl_df}
}
\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{attached_packages}},
\code{\link{startup_funs}}
}
\concept{startup_utils}
