% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign-add.R
\name{import}
\alias{import}
\title{Import}
\usage{
import(pgk, fun, overwrite = FALSE)
}
\arguments{
\item{fun}{String}

\item{overwrite}{Logical}

\item{pkg}{String}
}
\description{
Import a single function from a package
}
\examples{
\dontrun{
import("magrittr", "add")
}
}
