% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert}
\alias{insert}
\title{Insert}
\usage{
insert(x, positions, values)
}
\arguments{
\item{x}{A vector of values}

\item{positions}{Integer of positions of \code{x} to insert \code{values}}

\item{values}{A vector of values to insert into \code{x}}
}
\value{
A vector with the intended values inserted
}
\description{
Insert values at a position
}
\examples{
insert(letters[1:5], c(2, 4), c("X", "Y"))

}
