% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqrs.R
\name{iqrs}
\alias{iqrs}
\title{Inter Quartile Ranges}
\usage{
iqrs(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector of values}

\item{na.rm}{Logical. If `TRUE`, any `NA` and `NaN`'s are removed from `x` before the median and quantiles are computed.}
}
\description{
Computes the IQR magnitude of a vector
}
\examples{
iqrs(stats::rchisq(100, 2))
}
