% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{is_dir}
\alias{is_dir}
\alias{is_file}
\title{Is File/Directory}
\usage{
is_dir(x)

is_file(x)
}
\arguments{
\item{x}{A vector of file paths}
}
\value{
A \code{logical} vector
}
\description{
Is the path a file/directory?
}
\details{
These are essentially taken from \code{\link[utils:file_test]{utils::file_test()}} for \code{op = '-d'} and
\code{op = -f} but separated.
}
