% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jordan.R
\docType{package}
\name{jordan}
\alias{jordan}
\alias{jordan-package}
\title{jordan}
\description{
Just Other R Data Analytic Nuggets.  This package contains a variety of functions and wrappers developed for personally and professional use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmbarbone/jordan}
  \item Report bugs at \url{https://github.com/jmbarbone/jordan/issues}
}

}
