% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksource.R
\name{ksource}
\alias{ksource}
\title{Source an Rmd file}
\usage{
ksource(file, ..., quiet = TRUE, change_directory = FALSE)
}
\arguments{
\item{file}{An Rmd file.}

\item{...}{Additional arguments passes to [knitr::purl()]}

\item{quiet}{Logical; Determines whether to apply silence to [knitr::purl()]}

\item{change_directory}{Logical; if TRUE, the R working directory is temporarily
changed to the directory containing file for evaluating}
}
\description{
Sources an Rmd file as a temp file
}
