% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{assign_label}
\alias{assign_label.default}
\alias{assign_label.data.frame}
\alias{get_labels}
\title{Dataframe labels}
\usage{
assign_label(x, ...)

\method{assign_label}{default}(x, label, ...)

\method{assign_label}{data.frame}(x, ...)

get_labels(x)
}
\arguments{
\item{x}{A vector of data.frame}

\item{...}{One or more unquoted expressed separated by commas}

\item{label}{A single length string of a label to be assigned}
}
\description{
Assign labels to a vector or data.frame.
}
\examples{
## Best when run with RStudio

labs <- assign_label(iris,
                     Sepal.Length = "cms",
                     Sepal.Width  = "cms",
                     Petal.Length = "cms",
                     Petal.Width  = "cms",
                     Species      = "Iris ...")
# View(labs)
}
