% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_check}
\alias{length_check}
\alias{is_length0}
\alias{no_length}
\alias{has_length}
\title{Length checkers}
\usage{
is_length0(x)

no_length(x)

has_length(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\details{
\item{\code{is_length0}}{Not NULL but is length 0}
\item{\code{no_length}}{NULL or length 0}
\item{\code{has_length}}{Neither NULL or length 0}
}
}
\details{
Checks lengths
}
