% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{list_environments}
\alias{list_environments}
\alias{environments}
\alias{ls_all}
\alias{objects_all}
\title{List all environments and objects}
\usage{
environments()

ls_all(all.names = FALSE)

objects_all(all.names = FALSE)
}
\arguments{
\item{all.names}{a logical value.  If \code{TRUE}, all
    object names are returned.  If \code{FALSE}, names which begin with a
    \samp{.} are omitted.}
}
\value{
\itemize{
\item \code{\link[=environments]{environments()}}: Invisibly, a \code{character} vector of environment
names
}

\itemize{
\item \code{\link[=ls_all]{ls_all()}}, \code{\link[=objects_all]{objects_all()}}: A named list for each of the
environments the \code{\link[base:search]{base::search()}} path with all the objects found in that
environment
}
}
\description{
Functions to list out all environments and objects
}
\details{
\code{\link[=environments]{environments()}} is basically a printing wrapper for
\code{\link[base:search]{base::search()}}

\code{\link[=ls_all]{ls_all()}} and \code{\link[=objects_all]{objects_all()}} can be used retrieved all
objects from all environments in the \code{\link[base:search]{base::search()}} path, which may print
out a large result into the console.
}
