% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{loadAllNamespace}
\alias{loadAllNamespace}
\title{Get all from namespace}
\usage{
loadAllNamespace(namespace, warn.conflicts = TRUE)
}
\arguments{
\item{namespace}{The name of the namespace}

\item{warn.conflicts}{logical.  If \code{TRUE}, warnings are
    printed about \code{\link[base]{conflicts}} from attaching the database,
    unless that database contains an object \code{.conflicts.OK}.  A
    conflict is a function masking a function, or a non-function masking
    a non-function.
  }
}
\description{
Attaches package as "pseudo:\code{namespace}".
}
\details{
Loads all objects from a namespace
}
