% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{ls_ext}
\alias{ls_ext}
\alias{ls_dataframe}
\alias{ls_function}
\alias{ls_object}
\title{List Objects - extensions}
\usage{
ls_dataframe(pattern, all.names = FALSE)

ls_function(pattern, all.names = FALSE)

ls_object(pattern, all.names = FALSE)
}
\arguments{
\item{pattern}{an optional \link[base]{regular expression}.  Only names
    matching \code{pattern} are returned.  \code{\link{glob2rx}} can be
    used to convert wildcard patterns to regular expressions.}

\item{all.names}{a logical value.  If \code{TRUE}, all
    object names are returned.  If \code{FALSE}, names which begin with a
    \samp{.} are omitted.}
}
\value{
A \code{character} vector of names
}
\description{
List Objects - extensions
}
