% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mads.R
\name{mads}
\alias{mads}
\title{Median absolute deviations}
\usage{
mads(x, constant = 1.4826, ...)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{constant}{The scaling factor.}

\item{...}{additional arguments passes to `stats::mad`}
}
\description{
Computes the median absolute deviation
}
