% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-file.R
\name{make_sf}
\alias{make_sf}
\title{Make system file function}
\usage{
make_sf(package)
}
\arguments{
\item{package}{The name of the package}
}
\description{
Simple wrapper for package specific function for internal packages
}
