% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapply0}
\alias{mapply0}
\title{mapply with NULLs}
\usage{
mapply0(FUN, params = NULL, ...)
}
\arguments{
\item{FUN}{a function to call onto the parameters}

\item{params, ...}{params to pass to \code{FUN}}
}
\description{
mapply (basically) but with NULLs
}
\details{
This works similarly to \code{mapply()} expect it can take zero-length variables
}
\examples{
foo <- function(a, b, c) sum(a, b, c)
mapply(foo, a = 1:2, b = NULL, c = 3)
mapply0(foo, a = 1:2, b = NULL, c = 3)
# or pass as list
x <- list(a = 1:2, b = NULL, c = 3)
mapply0(foo, x)
}
