% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-cnd-conditions.R
\name{mark-cnd-conditions}
\alias{mark-cnd-conditions}
\alias{mark::append_warning}
\alias{mark::assign_labels_error}
\alias{mark::base_conversion_error}
\alias{mark::description_version_error}
\alias{mark::fct_expand_seq_error}
\alias{mark::import_error}
\alias{mark::internal_error}
\alias{mark::list2df_warning}
\alias{mark::match_arg_error}
\alias{mark::match_param_error}
\alias{mark::md5_condition}
\alias{mark::na_timezone_found_warning}
\alias{mark::not_available_error}
\alias{mark::numeric_datetime_tz_error}
\alias{mark::options_error}
\alias{mark::path_error}
\alias{mark::path_warning}
\alias{mark::reindex_error}
\alias{mark::reindex_warning}
\alias{mark::source_error}
\alias{mark::switch_error}
\alias{mark::timezone_not_found_error}
\alias{mark::view_labels_error}
\alias{append_warning}
\alias{assign_labels_error}
\alias{base_conversion_error}
\alias{description_version_error}
\alias{fct_expand_seq_error}
\alias{import_error}
\alias{internal_error}
\alias{list2df_warning}
\alias{match_arg_error}
\alias{match_param_error}
\alias{md5_condition}
\alias{na_timezone_found_warning}
\alias{not_available_error}
\alias{numeric_datetime_tz_error}
\alias{options_error}
\alias{path_error}
\alias{path_warning}
\alias{reindex_error}
\alias{reindex_warning}
\alias{source_error}
\alias{switch_error}
\alias{timezone_not_found_error}
\alias{view_labels_error}
\alias{mark:append_warning/warning}
\alias{mark:assign_labels_error/error}
\alias{mark:base_conversion_error/error}
\alias{mark:description_version_error/error}
\alias{mark:fct_expand_seq_error/error}
\alias{mark:import_error/error}
\alias{mark:internal_error/error}
\alias{mark:list2df_warning/warning}
\alias{mark:match_arg_error/error}
\alias{mark:match_param_error/error}
\alias{mark:md5_condition/condition}
\alias{mark:na_timezone_found_warning/warning}
\alias{mark:not_available_error/error}
\alias{mark:numeric_datetime_tz_error/error}
\alias{mark:options_error/error}
\alias{mark:path_error/error}
\alias{mark:path_warning/warning}
\alias{mark:reindex_error/error}
\alias{mark:reindex_warning/warning}
\alias{mark:source_error/error}
\alias{mark:switch_error/condition}
\alias{mark:timezone_not_found_error/error}
\alias{mark:view_labels_error/error}
\title{Conditions for \code{mark}}
\description{
Conditions for \code{mark}
}
\details{
The following conditions are defined in the \code{{mark}} package.
}
\section{\code{{cnd}}}{

These conditions are made with the \code{\link[cnd:cnd-package]{\{cnd\}}} package though
the use of \code{\link[cnd:condition]{cnd::condition()}}.
}

\section{\code{{mark}} conditions}{


\subsection{\code{mark:append_warning/warning}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::append_warning}}
\item{type}{\strong{warning}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:assign_labels_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::assign_labels_error}}
\item{type}{\strong{error}}
}
\code{assign_labels_error}

\strong{Columns not found}
You can set \code{.missing} to \code{warn} to get a warning instead of an error, or \code{skip} to silently skip those labels.

\strong{Malformed labels}
If passing labels as \code{.ls}, \code{...} must be empty. Columns in \code{...} must be entered as \code{name = value}; \code{column = label}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# instead of this:
assign_labels(df, a = 'AAA', .ls = list(b = 'BBB'))

# do this:
assign_labels(df, a = 'AAA', b = 'BBB')
# or this:
assign_labels(df, .ls = list(a = 'AAA', b = 'BBB'))
}\if{html}{\out{</div>}}

Labels must not be null; to remove labels, use \code{remove_labels()}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# instead of this
assign_labels(df, a = 'AAA', b = NULL)

# do this
df <- assign_labels(df, a = 'AAA')
df <- remove_labels(df, 'b')
}\if{html}{\out{</div>}}

}
\subsection{\code{mark:base_conversion_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::base_conversion_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:description_version_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::description_version_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:fct_expand_seq_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::fct_expand_seq_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:import_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::import_error}}
\item{type}{\strong{error}}
}
The object you are trying to import has already been assigned in the environment you are importing to.  Use the \code{overwrite} option to replace the object.

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# instead of
foo <- NULL
import('package', 'foo')

# do this
foo <- NULL
import('package', 'foo', overwrite = TRUE)
}\if{html}{\out{</div>}}

}
\subsection{\code{mark:internal_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::internal_error}}
\item{type}{\strong{error}}
}
Generic error to capture internal \code{{mark}} errors.  If any of these are encountered, please report an issue at \url{https://github.com/jmbarbone/mark/issues}
}
\subsection{\code{mark:list2df_warning/warning}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::list2df_warning}}
\item{type}{\strong{warning}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:match_arg_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::match_arg_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:match_param_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::match_param_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:md5_condition/condition}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::md5_condition}}
\item{type}{\strong{condition}}
}
Produces messages on md5 checks when \code{file_copy_md5(quiet = FALSE)}.
The message will indicate whether the file was new, or if the md5 hash was the same or different.
When \code{quiet = TRUE}, no messages will be produced.
}
\subsection{\code{mark:na_timezone_found_warning/warning}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::na_timezone_found_warning}}
\item{type}{\strong{warning}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:not_available_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::not_available_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:numeric_datetime_tz_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::numeric_datetime_tz_error}}
\item{type}{\strong{error}}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{# Instead of this:
diff_time(100, 200, tz = NULL)

# do this:
diff_time(100, 200, tz = 'America/New_York')

# or:
diff_time(100, 200, tz = 0)
}\if{html}{\out{</div>}}

}
\subsection{\code{mark:options_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::options_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:path_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::path_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:path_warning/warning}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::path_warning}}
\item{type}{\strong{warning}}
}
File creation cannot be performed when the path is an existing directory
}
\subsection{\code{mark:reindex_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::reindex_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:reindex_warning/warning}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::reindex_warning}}
\item{type}{\strong{warning}}
}
NA values in index may cause errors with expansion

\code{reindex()} will not match on NA values but instead will return a row of NAs

\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- data.frame(index = c(1:2, NA, 4:5), values = letters[1:5])
reindex(x, 'index', c(1, 2, 5))
#>   index values
#> 1     1      a
#> 2     2      b
#> 5     5      e

reindex(x, 'index', c(1, 2, 5, NA))
#>      index values
#> 1        1      a
#> 2        2      b
#> 5        5      e
#> <NA>    NA   <NA>
}\if{html}{\out{</div>}}

}
\subsection{\code{mark:source_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::source_error}}
\item{type}{\strong{error}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:switch_error/condition}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::switch_error}}
\item{type}{\strong{condition}}
}
\emph{no help documentation provided}
}
\subsection{\code{mark:timezone_not_found_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::timezone_not_found_error}}
\item{type}{\strong{error}}
}
When using a string for a timezone, this value must be found within \code{OlsonNames()}
}
\subsection{\code{mark:view_labels_error/error}}{
\describe{
\item{package}{\code{{mark}}}
\item{class}{\code{mark::view_labels_error}}
\item{type}{\strong{error}}
}
This may be because you are using RStudio :  https://community.rstudio.com/t/view-is-causing-an-error/75297/4 You can try :  \code{utils::View(get_labels(x), title = "Labels")}
}
}

\seealso{
\link[cnd:cnd-package]{cnd::cnd-package} \link[cnd:condition]{cnd::condition}
}
\keyword{internal}
