% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-list.R
\name{merge_list}
\alias{merge_list}
\title{Merge lists}
\usage{
merge_list(x, y, keep = c("x", "y"))
}
\arguments{
\item{x, y}{Lists to merge}

\item{keep}{When matching names are found, from which object should the values be
retained; \code{"x"} retains values from \code{x}, \code{"y"} retains values from \code{y}.}
}
\description{
Merge lists with different or intersecting names
}
\examples{
x <- list(a = 1, b = 2, c = NULL)
y <- list(a = 2, c = 3)
# compared to:
utils::modifyList(x, y)
merge_list(x, y)
merge_list(x, y, keep = "y")
}
