% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{muffle}
\alias{muffle}
\alias{wuffle}
\title{Muffle}
\usage{
muffle(expr, ...)

wuffle(expr, ...)
}
\arguments{
\item{expr}{An expression to be evaluated}

\item{...}{Additional arguments passed to \code{\link[base:message]{base::suppressMessages()}} or
\code{\link[base:warning]{base::suppressWarnings()}}}
}
\value{
The result of \code{expr}
}
\description{
Suppress messages and warnings
}
\details{
\code{muffle()} and \code{wuffle()} are aliases for \code{\link[base:message]{base::suppressMessages()}}
and \code{\link[base:warning]{base::suppressWarnings()}}, respectively, except the names are shorter
and therefore quicker to write.
}
