% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-at.R
\name{NA_at}
\alias{NA_at}
\alias{NA_if}
\title{NA at positions}
\usage{
NA_at(x, y, ...)

NA_if(x, y, ...)
}
\arguments{
\item{x}{A vector of values}

\item{y}{\code{NA_at()}: a vector of suitable index values.  \code{NA_if()}: a vector
of values to \code{match()} on.  Either can also accepted a function
which takes the vector \code{x} as its first argument and returns a valid \code{y}}

\item{...}{Additional values passed to \code{y} (if \code{y} is a function)}
}
\description{
Converts select elements of a vector into \code{NA}s

This is how the end results are
\itemize{
\item \code{NA_at} expects suitable index values (\code{x[y] <- NA})
\item \code{NA_if} expects some values to match on (\code{x[x \%in\% y] <- NA})
}
}
\details{
Convert specific values to NA
}
\seealso{
Inspired by \code{dplyr::na_if}
}
