% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{na_cols}
\alias{na_cols}
\alias{select_na_cols}
\alias{remove_na_cols}
\alias{is_na_cols}
\title{Selecting NA columns}
\usage{
select_na_cols(x)

remove_na_cols(x)

is_na_cols(x, names = TRUE)
}
\arguments{
\item{x}{A data.frame}
}
\description{
Select or remove columns that are entirely NA
}
