% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{norm_path}
\alias{norm_path}
\alias{file_path}
\alias{user_file}
\title{Normalize paths}
\usage{
norm_path(x = ".", check = FALSE, remove = check)

file_path(..., check = FALSE, remove = check)

user_file(..., check = FALSE, remove = check)
}
\arguments{
\item{x}{A character vector of paths}

\item{check}{Logical, if TRUE will check if the path exists and output a
warning if it does not.}

\item{remove}{Logical, if TRUE will remove paths that are not found}

\item{...}{Character vectors for creating a path}
}
\value{
A vector of full file paths
}
\description{
Normalize and check a vector of paths
}
