% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\alias{normalize.data.frame}
\title{Normalize values}
\usage{
normalize(x, range = base::range(x, na.rm = TRUE), bounds = 0:1)

\method{normalize}{default}(x, range = base::range(x, na.rm = TRUE), bounds = 0:1)

\method{normalize}{data.frame}(x, range = base::range(x, na.rm = TRUE), bounds = 0:1)
}
\arguments{
\item{x}{An object that is (coercible to) \code{double}}

\item{range}{The range of possible values of \code{x} as a vector of length \code{2}.
Defaults to \code{range(x, na.rm = TRUE)}}

\item{bounds}{The new boundaries for the normalized value as a vector of
length \code{2}.  Defaults to \code{c(0, 1)}.}
}
\value{
\code{x} with transformed values where \code{range} values are transformed to
\code{bounds}.
}
\description{
Normalizes values based on possible range and new bounds
}
\examples{
x <- c(0.23, 0.32, 0.12, 0.61, 0.26, 0.24, 0.23, 0.32, 0.29, 0.27)
data.frame(
  x = normalize(x),
  v = normalize(x, range = c(0, 2)),
  b = normalize(x, bounds = c(0, 10)),
  vb = normalize(x, range = c(0, 2), bounds = c(0, 10))
)

# maintains matrix
mat <- structure(c(0.24, 0.92, 0.05, 0.37, 0.19, 0.69, 0.43, 0.22, 0.85,
0.73, 0.89, 0.68, 0.57, 0.89, 0.61, 0.98, 0.75, 0.37, 0.24, 0.24,
0.34, 0.8, 0.25, 0.46, 0.03, 0.71, 0.79, 0.56, 0.83, 0.97), dim = c(10L, 3L))

mat
normalize(mat, bounds = c(-1, 1))
normalize(as.data.frame(mat), bounds = c(-1, 1))
}
