% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-available.R
\name{not_available}
\alias{not_available}
\alias{set_not_available}
\alias{NA_Date_}
\alias{NA_POSIXct_}
\alias{NA_POSIXlt_}
\title{Make not available}
\usage{
not_available(type = "logical", length = 0L)

set_not_available(type, value)

NA_Date_

NA_POSIXct_

NA_POSIXlt_
}
\arguments{
\item{type}{Type of NA (see details)}

\item{length}{Length of the vector}

\item{value}{A value to return in \code{not_available()}}
}
\value{
A vector of \code{NA} values
}
\description{
Create NA vectors
}
\details{
If length is a text it will search for an appropriate match.
}
\examples{
x <- not_available("Date", 3)
x
class(x)

}
