% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\alias{note<-}
\alias{print.jordan_note}
\title{Append a note to an object}
\usage{
note(x) <- value

note(x)

\method{print}{jordan_note}(x, ...)
}
\arguments{
\item{x}{An object}

\item{value}{The note to attach}

\item{...}{Additional arguments passed from methods (not used)}
}
\description{
An alternative to the [base::comment()].
}
\details{
When the note is assigned to an object a new class will be added,
  `jordan_note`, so that a `print` function can call an S3 method.  The print
  for this can be adjusted for it's width by using the option
  `jordan.note.width` which defaults to the option `width` when not set.

The method of _printing_ can also be adjusted with the `jordan.note.fun`
  option in which a function can be assigned.  This function will then be
  applied to the note at print.

The type of object assigned to the note is not restricted, so user beware
  of odd prints or additional features added to the notes fun.
}
