% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_default}
\alias{null_default}
\alias{\%||\%}
\title{Default value for NULL}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If \code{x} is \code{NULL} returns \code{y}; otherwise \code{x}}
}
\description{
Replace if \code{NULL}
}
\details{
A mostly copy of \code{rlang}'s \verb{\%||\%} except does not use \code{\link[rlang:type-predicates]{rlang::is_null()}},
which, currently, calls the same primitive \code{is.null} function as
\code{\link[base:NULL]{base::is.null()}}.
}
