% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{odds2d}
\alias{odds2d}
\title{Odds ratio to Cohen's D}
\usage{
odds2d(odds, var = FALSE, na.rm = FALSE)
}
\arguments{
\item{odds}{An odds ratio}

\item{var}{Logical, if \code{TRUE} calculates the variances of d}

\item{na.rm}{Logical, if TRUE \code{NA} values are removed
(only needed for \code{var = TRUE})}
}
\description{
Odds ratio to Cohen's D
}
\examples{
odds2d(c(1.1, .001, 2, NA)) # cohen's D for each
odds2d(c(1.1, .001, 2, NA), var = TRUE)
odds2d(c(1.1, .001, 2, NA), var = TRUE, na.rm = TRUE)

}
