% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Odds ratio}
\usage{
odds_ratio(a, b, c, d, type = "hits_misses")
}
\arguments{
\item{a, b, c, d}{Values in the contingency table.}

\item{type}{Type to use}
}
\value{
The odds ratio
}
\description{
Returns the odds ratio without returning infinity
}
