% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{omit_na}
\alias{omit_na}
\title{Omit NA values}
\usage{
omit_na(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
\code{x} which \code{NA} values removes and two attributes of \code{integers}: \code{na}
which is the position of \code{NA} values, and \code{valid} for the position of
non-\code{NA} values; empty positions reported as \code{integer(0)}
}
\description{
Omit NA values
}
\examples{
# Like stats::na.omit but always provides
x <- letters[1:5]
omit_na(x)
x[c(3, 5)] <- NA
omit_na(x)

}
