% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{open_file}
\alias{open_file}
\alias{shell_exec}
\alias{list_files}
\title{Open a file using windows file associations}
\usage{
open_file(x)

shell_exec(x)

list_files(x, pattern = NULL, ignore_case = FALSE, all = FALSE)
}
\arguments{
\item{x}{A vector of paths}

\item{pattern}{an optional \link[base]{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{ignore_case}{logical. Should pattern-matching be case-insensitive?}

\item{all}{a logical value. If FALSE, only the names of visible files are
returned (following Unix-style visibility, that is files whose name does
not start with a dot). If TRUE, all file names will be returned.}
}
\description{
Opens the given files(s)
}
\details{
`open_file` is an alternative to [base::shell.exec()] that can take take
  multiple files.
`list_files` is mostly a wrapper for [base::list.files()] with preferred
  defaults.
}
