% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc-utils.R
\name{optimal_threshold}
\alias{optimal_threshold}
\title{Optimal threshold from pROC mod}
\usage{
optimal_threshold(mod, method = c("youden", "top_left"), ...)
}
\arguments{
\item{mod}{A roc mod from the pROC package}

\item{method}{Method to determine optimal threshold}

\item{...}{Additional arguments passed to [pROC::ci.thresholds()]}
}
\description{
Find the optimal threshold from a pROC mod
}
\examples{
x <- pROC::aSAH
mod <- pROC::roc(x$outcome, x$s100b, levels=c("Good", "Poor"))

optimal_threshold(mod)
}
