% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-values.R
\name{p_values}
\alias{p_values}
\alias{p_round}
\alias{p_value_sig}
\title{Round p-value}
\usage{
p_round(x, n = 3, sig = n)

p_value_sig(x, one = 0.001, two = 0.01, three = 0.05, four = 0.1)
}
\arguments{
\item{x}{A vector of p-values}

\item{n}{Number of digits to round by (if NULL - no rounding occurs)}

\item{sig}{Number of significant figures (if NULL - not used)}

\item{one, two, three, four}{The significance thresholds (defaults to .001, .01, .05, and .10).}
}
\description{
Rounds a p-value by decimal places and reports with sig figs
}
\examples{
x <- stats::pchisq(abs(runif(25)), 3)
print(data.frame(x = x,
                 p = p_round(x),
                 sigs = p_value_sig(x)),
      digits = 3,
      right = FALSE)
}
