% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_c}
\alias{paste_c}
\title{Paste combine}
\usage{
paste_c(x, y, collate = TRUE, sep = "")
}
\arguments{
\item{x, y}{Vectors to paste and combine}

\item{collate}{Logical; \code{TRUE} prints out combinations in order of vector
\code{x} elements then \code{y}; otherwise reversed (see examples)}

\item{sep}{A character string to separate terms}
}
\description{
Paste and combine
}
\examples{
x <- letters[1:5]
y <- 1:3
paste_c(x, y)
paste_c(x, y, sep = "_")
paste_c(x, y, collate = FALSE)
}
