% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_norm}
\alias{path_norm}
\title{Normalize a path}
\usage{
path_norm(x, check = FALSE, remove = check)
}
\arguments{
\item{x}{A vector of paths}

\item{check}{Logical, if TRUE will check if the path exists and output a
warning if it does not.}

\item{remove}{Logical, if TRUE will remove paths that are not found}
}
\description{
Normalize and check a vector of paths
}
