% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{print.mark_bib_df}
\alias{print.mark_bib_df}
\title{Print bib data frame}
\usage{
\method{print}{mark_bib_df}(x, list = FALSE, ...)
}
\arguments{
\item{x}{The \code{mark_bib_df} object}

\item{list}{If \code{TRUE} will print as a list rather than the \code{data.frame}}

\item{...}{Additional arguments passed to methods}
}
\value{
\code{x}, invisibly, called for its side effects
}
\description{
Print bib dataframe, or as list
}
