% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo-id.R
\name{print.pseudo_id}
\alias{print.pseudo_id}
\title{Print \code{pseudo_id}}
\usage{
\method{print}{pseudo_id}(x, ..., all = FALSE)
}
\arguments{
\item{x}{An object of class \link{pseudo_id}}

\item{...}{Not implemented}

\item{all}{if \code{TRUE} will print all uniques.  This is not recommend for many
uniques as it will crowd the console output}
}
\value{
\code{x}, invisibly.  Called for its side effects.
}
\description{
Print \code{pseudo_id}
}
\seealso{
\code{\link[=pseudo_id]{pseudo_id()}}
}
