% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc-utils.R
\name{pROC_quick_plot}
\alias{pROC_quick_plot}
\title{Plots an ROC model}
\usage{
pROC_quick_plot(
  mod,
  thres_method = c("youden", "closest.topleft"),
  col = "blue",
  ...,
  boots = 0L
)
}
\arguments{
\item{mod}{An object with class "roc" (a ROC model made with \code{\link[pROC:roc]{pROC::roc()}})}

\item{thres_method}{The threshold method to print}

\item{col}{The color of the curve}

\item{...}{Additional arguments passed to \code{\link[pROC:plot.roc]{pROC::plot.roc()}}}

\item{boots}{Number of bootstrap replications to perform; if < 2L will not
perform any}
}
\description{
Quickly plots a model based on my own liking
}
\details{
When adding bootstrapped confidence intervals, the core function of
\code{pROC:::ci.sp.roc()} is replaced with a quicker version that uses parallel
processing to speed up the bootstraps and some other data manipuluation.
There will likely be a delay when running with bootstraps.
}
\examples{
x <- pROC::aSAH
mod <- pROC::roc(x$outcome, x$s100b, levels = c("Good", "Poor"))
pROC_quick_plot(mod)

\dontrun{
pROC_quick_plot(mod, boots = 100)
}
}
