% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib.R
\name{process_bib_dataframe}
\alias{process_bib_dataframe}
\title{Process bib values}
\usage{
process_bib_dataframe(categories, values, fields, keys)
}
\arguments{
\item{categories}{A list of categories}

\item{values}{A list of values}

\item{fields}{a Vector of fields}

\item{keys}{a Vector of keys}
}
\value{
A wide \code{data.frame} with explicit \code{NA}s
}
\description{
Generates a data frame of values from bibs
}
