% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo_id}
\alias{pseudo_id}
\title{Create an ID for a vector}
\usage{
pseudo_id(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Transforms a vector into an integer of IDs.
}
\examples{
\dontrun{
set.seed(42)
(x <- sample(letters, 10, TRUE))
(pid <- pseudo_id(x))
attr(pid, "uniques")[pid]
}
}
