% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{quick_df}
\alias{quick_df}
\title{Quick DF}
\usage{
quick_df(x)
}
\arguments{
\item{x}{A list}
}
\description{
This is a speedier implementation of \code{as.data.frame()} but does not provide
the same sort of checks. It should be used with caution.
}
