% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc-utils.R
\name{quick_plot_roc}
\alias{quick_plot_roc}
\title{Plots an ROC model}
\usage{
quick_plot_roc(
  mod,
  thres_method = c("youden", "closest.topleft"),
  col = "blue",
  ...
)
}
\arguments{
\item{mod}{A model from pROC}

\item{thres_method}{The threshold method to print}

\item{col}{The color of the line}

\item{...}{Additional arguments passed to [pROC::plot.roc()]}
}
\description{
Quickly plots a model based on my own likings
}
\examples{
x <- pROC::aSAH
mod <- pROC::roc(x$outcome, x$s100b, levels=c("Good", "Poor"))

quick_plot_roc(mod)
}
