% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bib.R
\name{read_bib}
\alias{read_bib}
\title{Read Bib file}
\usage{
read_bib(file, n = -1L, encoding = "UTF-8")
}
\arguments{
\item{file}{File or connection}

\item{n}{Integer, the maximum number of lines to read (passed to readLines)}

\item{encoding}{Assumed encoding of file (passed to readLines)}
}
\description{
Read a bib file into a data.frame
}
\details{
Inspired and partially credited to [bib2df::bib2df()] although this has no
  dependencies outside of base functions and much quicker.  This speed seems
  to come from removing `stringr` functions and simplifying a few *apply
  functions.
This will also include as many categories as possible from the entry.
}
\examples{
file <- "https://raw.githubusercontent.com/jmbarbone/bib-references/master/references.bib"
bibdf <- read_bib(file, n = 51L)
if (requireNamespace("tibble")) tibble::as_tibble(bibdf) else head(bibdf)

if (requireNamespace("bib2df") & requireNamespace("microbenchmark")) {
  file1 <- system.file("extdata", "bib2df_testfile_3.bib", package = "bib2df")
  microbenchmark::microbenchmark(
    read_bib(file1),
    bib2df::bib2df(file1)
  )
}
}
