% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{remove_null}
\alias{remove_null}
\title{Remove NULL}
\usage{
remove_null(x)
}
\arguments{
\item{x}{A list}
}
\value{
The list \code{x} without \code{NULL}
}
\description{
Remove NULL results from a list
}
\examples{
x <- list(a = letters[1:5], b = NULL, c = complex(3))
x
remove_null(x)
}
