% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_each.R
\name{rep_each}
\alias{rep_each}
\title{Repeat each}
\usage{
rep_each(.x, times = 1)
}
\arguments{
\item{.x}{A vector of values to repeat}

\item{times}{The number of times to repeat each value in in `.x`}
}
\value{
a `vector` with the values repeated in the order in which they were called
}
\description{
Repeats each value in order they are passed rather than
}
