% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{require_namespace}
\alias{require_namespace}
\title{Require namespace}
\usage{
require_namespace(namespace)
}
\arguments{
\item{namespace}{The name of a package/namespace}
}
\description{
A wrapped requireNamespace
}
\examples{
foo <- function() {
  require_namespace("bad_package")
  1
}

\dontrun{
require_namespace("bad_package")
foo()
}

}
\references{
http://r-pkgs.had.co.nz/description.html
}
