% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\alias{reverse.default}
\alias{reverse.matrix}
\alias{reverse.data.frame}
\title{Reverse an object}
\usage{
reverse(x, ...)

\method{reverse}{default}(x, ...)

\method{reverse}{matrix}(x, preserve_rownames = FALSE, ...)

\method{reverse}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments passed to methods}

\item{preserve_rownames}{Logical, if \code{TRUE} will not reset rownames}
}
\description{
Reverse an object
}
