% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse-log-trans.R
\name{reverse_log_trans}
\alias{reverse_log_trans}
\title{Reverse log transformation for ggplot}
\usage{
reverse_log_trans(base = exp(1))
}
\arguments{
\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}
}
\description{
Show a reversed order for a log transformed scale from ggplot.
Ripped from: https://stackoverflow.com/questions/11053899/how-to-get-a-reversed-log10-scale-in-ggplot2
}
\author{
Brian Diggs (https://stackoverflow.com/users/892313/brian-diggs)
}
