% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-by.R
\name{round_by}
\alias{round_by}
\title{Rounding by a specific interval.}
\usage{
round_by(x, by = 1, method = "round")
}
\arguments{
\item{x}{A number or vector to round.}

\item{by}{The number by which to round}

\item{method}{An option to explicitly specify automatic rounding, ceiling, or floor}
}
\description{
Rounds a number or vector of numbers by another
}
\examples{
round_by(sample(100, 100, replace = TRUE), by = 5)
round_by(seq(1, 20, .6), 2, "ceiling")
round_by(runif(10), .01, "floor")
round_by(runif(10) * 100, seq(10))
}
