% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_to}
\alias{round_to}
\title{Round to}
\usage{
round_to(x, anchors)
}
\arguments{
\item{x}{A vector of values}

\item{anchors}{A vector of anchor values}
}
\description{
Rounds a vector to the nearest in a set of anchors.
}
\examples{
x <- rpois(10, 1)
anchors <- c(0, 0.5, 1.5, 3)
data.frame(x, anchor = round_to(x, anchors))
}
