% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcing.R
\name{rscript}
\alias{rscript}
\title{Rscript}
\usage{
rscript(x, ops = NULL, args = NULL, ...)
}
\arguments{
\item{x}{An R file to run}

\item{ops}{A character vector of options (\code{"--"} is added to each)}

\item{args}{A character vector of other arguments to pass}

\item{...}{Additional arguments passed to \code{system2}}
}
\value{
A \code{character} vector of the result from calling \code{Rscript} via
\code{system2()}
}
\description{
Implements \code{Rscript} with \code{system2}
}
\seealso{
\link{source_to_env}
}
