% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcing.R
\name{save_source}
\alias{save_source}
\title{Save source}
\usage{
save_source(env = parent.frame(), file = mark_temp("Rds"), name = NULL)
}
\arguments{
\item{env}{The parent environment}

\item{file}{The file to save the environment to}

\item{name}{An optional name for the environment (mostly cosmetic)}
}
\value{
A \code{source_env}/\code{environment} object, created from \code{env}
}
\description{
Source a file and save as file
}
