% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_pooled.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Pooled standard deviation}
\usage{
sd_pooled(ns, ses, max = FALSE)
}
\arguments{
\item{ns}{A vector of N values.}

\item{ses}{A vector of standard errors.}

\item{max}{Logical value.  Whether to compute the max pooled deviation.}
}
\description{
Computes the pooled standard error
}
