% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setOption}
\alias{setOption}
\title{Set options}
\usage{
setOption(name, value)
}
\arguments{
\item{name}{The name of the option}

\item{value}{The value of the option}
}
\description{
Dynamically set options
}
\examples{
\dontrun{
opt_name <- "testing.option"
opt_value <- TRUE
setOption(opt_name, opt_value)
getOption(opt_name)
opt_value <- "Jordan"
setOption(opt_name, opt_value)
getOption(opt_name)
opt_value <- 10
setOption(opt_name, opt_value)
getOption(opt_name)
}
}
