% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-report.R
\name{simpleTimeReport}
\alias{simpleTimeReport}
\title{Time reports}
\usage{
simpleTimeReport(title = NULL, expr, envir = parent.frame())
}
\arguments{
\item{title}{The title to be printed}

\item{expr}{The expression to run}

\item{envir}{The environment from which to evaluate the \code{expr}}
}
\description{
Evaluate code and report on the time difference
}
\examples{
\dontrun{
simpleTimeReport("example", {
  print("1")
  Sys.sleep(1)
  warning("this is a warning")
  sample(1e6, 1e6, TRUE)
})
}
}
