% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-by.R
\name{sort_by}
\alias{sort_by}
\title{Sort by}
\usage{
sort_by(x, by, ...)
}
\arguments{
\item{x}{A vector}

\item{by}{Another vector}

\item{...}{Additional arguments passed to \code{base::order()}}
}
\value{
The values of \code{x}, resorted
}
\description{
Sort an object by another object
}
\examples{
l3 <- letters[1:3]
sort_by(l3, c(3, 2, 1))
# make a factor object with the reversed order
f <- factor(l3, levels = rev(l3))
sort_by(f, l3)
sort_by(1:3, rev(l3))

}
