% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{sort_names}
\alias{sort_names}
\title{Sort by names}
\usage{
sort_names(x, numeric = FALSE)
}
\arguments{
\item{x}{A named vector of values}

\item{numeric}{If \code{TRUE} will try to coerce to numeric}
}
\value{
\code{x} sorted by its \code{names()}
}
\description{
Sort a vector by it's name
}
