% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-r-dir.R
\name{source_dir_r}
\alias{source_dir_r}
\title{Source file from directory}
\usage{
source_dir_r(dir, quiet = F)
}
\arguments{
\item{dir}{The location of your R scripts}

\item{quiet}{Logical.  Whether to print out a message for each file.}
}
\description{
Walk through files in a directory and output them
}
\examples{
\dontrun{
source_dir_r("C:/Users/jbarbone/GitHub/easyr/R/")
}
}
