% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-r-dir.R
\name{source_files}
\alias{source_files}
\alias{source_r_dir}
\alias{source_dir_r}
\alias{source_r_file}
\title{Source file from directory}
\usage{
source_r_dir(dir, echo = FALSE, quiet = FALSE, ...)

source_dir_r(dir, echo = FALSE, quiet = FALSE, ...)

source_r_file(path, echo = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{dir}{The location of your R scripts}

\item{echo}{logical; if \code{TRUE}, each expression is printed
    after parsing, before evaluation.}

\item{quiet}{Logical.  Whether to print out a message for each file.}

\item{...}{Additional arguments passed to \code{\link[base:source]{base::source()}}}

\item{path}{The location of the R file.}
}
\description{
Walk through files in a directory and output them.
Files are sources in order of names
}
\examples{
\dontrun{
source_dir_r("C:/Users/jbarbone/GitHub/test")
}
}
