% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcing.R
\name{source_to_env}
\alias{source_to_env}
\title{Source to environment}
\usage{
source_to_env(x, ops = NULL)
}
\arguments{
\item{x}{An R script}

\item{ops}{Options to be passed to \link{rscript}}
}
\value{
Invisibly, and environment variable of the objects/results created
from \code{x}
}
\description{
Source an R script to an environment
}
