% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-report.R
\name{split_expression}
\alias{split_expression}
\title{Split an expression}
\usage{
split_expression(expr)
}
\arguments{
\item{expr}{An expression to split}
}
\description{
Can be useful when passing an expression with curly braces to separate into
separate expressions to evaluate (and report) out one by one
}
\details{
Split an expression into multiple expressions
}
\examples{
\dontrun{
ss <- split_expression({
  mean(1:10)
  sample(
    letters,
    5,
    TRUE
  )
  two <- 2
  two * 2
})

for (i in ss) {
  str(charexpr(i))
}
}
}
