% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\docType{data}
\name{startup_funs}
\alias{startup_funs}
\alias{.LoadFunctionsFromJordan}
\alias{.RunDefaultFunctionsFromJordan}
\alias{.git_branch_prompt}
\alias{.load_pipe}
\title{Start up functions}
\format{
An object of class \code{character} of length 12.
}
\usage{
.LoadFunctionsFromJordan(env = parent.frame())

.RunDefaultFunctionsFromJordan(env = parent.frame())

startup_funs

.git_branch_prompt()

.load_pipe(env = parent.frame())
}
\arguments{
\item{env}{The environment to save objects}
}
\description{
Functions that try to assist with start ups
}
\details{
These functions can be used to create an environment for startup.
}
\section{Start up}{

For more information on start up:
\itemize{
\item https://rstats.wtf/r-startup.html
\item \code{?Startup}
}
}

\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{attached_packages}},
\code{\link{ht}()}
}
\concept{startup_utils}
\keyword{datasets}
