% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sterr.R
\name{sterr}
\alias{sterr}
\title{Standard error}
\usage{
sterr(x, na.rm = F)
}
\arguments{
\item{x}{A vector or data.frame}

\item{na.rm}{Passed to \code{stats::sd()}.}
}
\value{
A vector
}
\description{
Returns the standard error of a vector or columns of a data.frame
}
