% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_extract_date}
\alias{str_extract_date}
\title{Extract date from string}
\usage{
str_extract_date(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{A character vector}

\item{format}{A date format to find}
}
\value{
A date (if found) or NA
}
\description{
Extract date from string
}
\examples{
str_extract_date("This is a file name 2020-02-21.csv")
str_extract_date(c("This is a file name 2020-02-21.csv", "No date"))
str_extract_date("Last saved 17 December 2019", format = "\%d \%B \%Y")
}
