% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch-params.R
\name{switch_params}
\alias{switch_params}
\title{Switch with a list of parameters}
\usage{
switch_params(x, params)
}
\arguments{
\item{x}{A single value}

\item{params}{A list of cases}
}
\description{
Switch with a list of params
}
\references{
Original: https://stackoverflow.com/a/32835930/12126576
}
