% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{t_df}
\alias{t_df}
\title{Data frame transpose}
\usage{
t_df(x, id = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{id}{No longer used}
}
\description{
Transposes a data.frame as a data.frame
}
\examples{
x <- data.frame(col_a = Sys.Date() + 1:5, col_b = letters[1:5], col_c = 1:5)
t_df(x)
}
