% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilz.R
\name{that}
\alias{that}
\title{That}
\usage{
that(x, arr.ind = FALSE, useNames = TRUE)
}
\arguments{
\item{x}{a \code{\link[base]{logical}} vector or array.  \code{\link[base]{NA}}s
   are allowed and omitted (treated as if \code{FALSE}).}

\item{arr.ind}{logical; should \bold{arr}ay \bold{ind}ices be returned
   when \code{x} is an array?}

\item{useNames}{logical indicating if the value of \code{arrayInd()}
   should have (non-null) dimnames at all.}
}
\description{
Grammatical correctness
}
\details{
See `fortunes::fortune(175)`.
}
\seealso{
[base::which()]
}
