% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-time.R
\name{time_inherits}
\alias{time_inherits}
\alias{is_POSIXlt}
\alias{is_POSIXct}
\alias{is_diff_time}
\title{Time inherits}
\usage{
is_POSIXlt(x)

is_POSIXct(x)

is_diff_time(x)
}
\arguments{
\item{x}{An object}
}
\description{
\code{inherits(x, ...)} wrappers
}
\seealso{
\link{diff_time}
}
