% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-row-names.R
\name{to_row_names}
\alias{to_row_names}
\title{To row names}
\usage{
to_row_names(data, row_names = 1L)
}
\arguments{
\item{data}{A data.frame}

\item{row_names}{The numeric position of the column.}
}
\description{
Converts a column to row names
}
\examples{

x <- data.frame(
  a = 1:4,
  b = letters[1:4]
)

to_row_names(x)
to_row_names(x, "b")
}
