% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksource.R
\name{try_source}
\alias{try_source}
\alias{try_ksource}
\title{Try to source a file}
\usage{
try_source(file, ..., change_directory = FALSE)

try_ksource(file, ..., change_directory = FALSE)
}
\arguments{
\item{file}{An R or Rmd file}

\item{...}{Additional arguments passes to [knitr::purl()]}

\item{change_directory}{Logical; if TRUE, the R working directory is temporarily
changed to the directory containing file for evaluating}
}
\description{
This will output an error message rather than completely preventing the execution.
This can be useful for when a script calls on multiple, independent files to be sourced
and a single failure shouldn't prevent the entire run to fail as well.
}
