% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tryn.R
\name{tryn}
\alias{tryn}
\title{Try an expression a set number of times}
\usage{
tryn(expr, n = 10, silent = TRUE)
}
\arguments{
\item{expr}{expression to evaluate}

\item{n}{number of attempts until error}

\item{silent}{whether to suppress warnings}
}
\value{
result of \code{expr}
}
\description{
Try an expression a set number of times
}
\examples{
foo <- function() stop("I added an error")
try(tryn(n = 10, foo()))
}
