% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{tukey_coef}
\alias{tukey_coef}
\title{Tukey's coefficient}
\usage{
tukey_coef(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
The coefficient of the distance from Q1 or Q3 by IQR values
}
\description{
Returns the coefficient for Tukey's boxplot
}
\examples{
tukey_coef(0:20)
}
