% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist.R
\name{unlist0}
\alias{unlist0}
\title{Unlist}
\usage{
unlist0(x)
}
\arguments{
\item{x}{A list}
}
\description{
Unlist without unique names
}
\examples{
x <- list(
  a = 1:3,
  b = 2,
  c = 2:4
)

unlist(x)   # names: a1 a2 a3  b c1 c2 c3
unlist0(x)  # names: a a a  b c c c
}
