% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{use_author}
\alias{use_author}
\title{Add author to DESCRIPTION}
\usage{
use_author(author_info = find_author())
}
\arguments{
\item{author_info}{Author information as a named list}
}
\value{
None, called for side effects
}
\description{
Adds author to description
}
\details{
Only valid for a single author.
}
