% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{user_file}
\alias{user_file}
\title{File path in user directory}
\usage{
user_file(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{character vectors.  Long vectors are not supported.}

\item{fsep}{the path separator to use (assumed to be ASCII).}
}
\description{
A wrapper for creating a file name under your R_USER directory.
}
\examples{
\dontrun{
data(iris)
file_path <- user_file("outputs/tests.csv")
write.csv(iris, file = file_path)
}
}
\seealso{
[base::file.path()]
}
