% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{vap}
\alias{vap}
\alias{vap_int}
\alias{vap_dbl}
\alias{vap_chr}
\alias{vap_lgl}
\alias{vap_cplx}
\alias{vap_date}
\title{Vaps!}
\usage{
vap_int(.x, .f, ..., .nm = FALSE)

vap_dbl(.x, .f, ..., .nm = FALSE)

vap_chr(.x, .f, ..., .nm = FALSE)

vap_lgl(.x, .f, ..., .nm = FALSE)

vap_cplx(.x, .f, ..., .nm = FALSE)

vap_date(.x, .f, ..., .nm = FALSE)
}
\arguments{
\item{.x}{A vector of values}

\item{.f}{A function to apply to each element in vector \code{.x}}

\item{...}{Additional arguments passed to \code{.f}}

\item{.nm}{Logical, if \code{TRUE} returns names of \code{.x} (Note: If \code{.x} does not
have any names, they will be set to the values)}
}
\value{
A vector of type matching the intended value in the function name.
}
\description{
Wrappers for vapply
}
\details{
These are simply wrappers for \code{\link[base:vapply]{base::vapply()}} to shorten lines.

Each function is designed to use specific vector types:

\describe{
\item{\code{\link[=vap_int]{vap_int()}}}{integer}
\item{\code{\link[=vap_dbl]{vap_dbl()}}}{double}
\item{\code{\link[=vap_chr]{vap_chr()}}}{character}
\item{\code{\link[=vap_lgl]{vap_lgl()}}}{logical}
\item{\code{\link[=vap_cplx]{vap_cplx()}}}{complex}
\item{\code{\link[=vap_date]{vap_date()}}}{Date}
}
}
\seealso{
\code{\link[base:vapply]{base::vapply()}}
}
