% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2df.R
\name{vector2df}
\alias{vector2df}
\title{Vector to data.frame}
\usage{
vector2df(.vector, show_NA = F)
}
\arguments{
\item{.vector}{A vector of values.}

\item{show_NA}{Logical.}
}
\description{
Transforms a vector (named) to a data.frame
}
