% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{with_par}
\alias{with_par}
\title{Temporary plotting}
\usage{
with_par(..., ops = NULL)
}
\arguments{
\item{...}{Code to be evaluated}

\item{ops}{A named list to be passed to \code{\link[graphics:par]{graphics::par()}}}
}
\value{
Invisibly, the result of \code{...}
}
\description{
Reset par() after running
}
\examples{
with_par(
  plot(lm(Sepal.Length ~ Sepal.Width, data = iris)),
  plot(lm(Petal.Length ~ Petal.Width, data = iris)),
  ops = list(mfrow = c(2, 4))
)
}
