% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{within}
\alias{within}
\alias{between}
\alias{between_more}
\title{within boundaries}
\usage{
between_more(x, left, right, type = c("gele", "gel", "gle", "gl"))

within(x, left = NULL, right = NULL, bounds = c("[]", "[)", "(]", "()"))
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values.  For \code{\link[=within]{within()}}, when \code{NULL} no
comparison is made for that boundary.  When both are \code{NULL}, \code{x} is just
returned.}

\item{type}{Abbreviation for the evaluation of \code{left} on \code{right} (see
details)}

\item{bounds}{Boundaries for comparisons of \code{left} and \code{right} (see details)}
}
\value{
A logical vector
}
\description{
Compare a vector within (between) other values
}
\details{
\code{type}, \code{bounds} can be one of the below:

\describe{
\item{\code{"g"}, \code{(}}{is greater than (\code{>})}
\item{\code{"ge"}, \code{[}}{greater than or equal to (\code{>=})}
\item{\code{"l"}, \verb{)}}{less than (\code{<})}
\item{\code{"le"}, \verb{]}}{less than or equal to (\code{<=})}
}

Note: [mark::between_more()] may be deprecated in the future in favor of
just [mark::within()]

[\verb{\}\{greater than or equal to (}>=\verb{)\} \\item\{}"l"\verb{, })\verb{\}\{less than (}<\verb{)\} \\item\{}"le"\verb{, }]: R:\%60\%7D\%7Bgreater\%20than\%20or\%20equal\%20to\%20(\%60\%3E=\%60)\%7D\%0A\%20\%5C\%5Citem\%7B\%60\%22l\%22\%60,\%20\%60)\%60\%7D\%7Bless\%20than\%20(\%60\%3C\%60)\%7D\%0A\%20\%5C\%5Citem\%7B\%60\%22le\%22\%60,\%20\%60
[mark::between_more()]: R:mark::between_more()
[mark::within()]: R:mark::within()
}
\examples{

between_more(2:10, 2, 10, "gl")
within(2:10, 2, bounds = "()")
between_more(10, 2, 10, "gle")
within(2:10, bounds = "(]")
within(1:5, c(3, 3, 2, 2, 1), 5)
}
