% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{within_call}
\alias{within_call}
\alias{within_fun}
\alias{outer_call}
\alias{outer_fun}
\title{Function within}
\usage{
within_call()

within_fun()

outer_call(n = 0)

outer_fun(n = 0)
}
\arguments{
\item{n}{The number of calls to move out from}
}
\value{
The string of the call/function
}
\description{
Returns the function call you are within
}
