% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_file_ms}
\alias{write_file_ms}
\title{Write file with md5 hash check}
\usage{
write_file_ms(
  x,
  path = NULL,
  method = NULL,
  overwrite = NA,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{x}{An object to write to file}

\item{path}{A character vector of one or more paths.}

\item{method}{The method of saving the file.  When \code{NULL}, the method is
determined by the file extension.}

\item{overwrite}{When \code{NA}, only saves if the md5 hashes do not match.
Otherwise, see \code{\link[fs:copy]{fs::file_copy()}}.}

\item{quiet}{When \code{TRUE}, suppresses messages from md5 checks.}

\item{...}{Additional arguments passed to the write function.}

\item{file}{The file or connection to write to (dependent on part by method)}
}
\value{
See \code{\link[=file_copy_ms]{file_copy_ms()}}.  When \code{file} is a connection, the returns
nothing and writes the file to the \code{stdout()}
}
\description{
Write file with md5 hash check
}
\examples{
df <- data.frame(a = 1, b = 2)

# just writes to stdout()
write_file_ms(df)

temp <- tempfile()
# new
write_file_ms(df, temp)

# same
write_file_ms(df, temp)

# changed
df$c <- 3
write_file_ms(df, temp)

fs::file_delete(temp)
}
