% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{as_ordered}
\alias{as_ordered}
\alias{as_ordered.default}
\alias{as_ordered.factor}
\alias{as_ordered.ordered}
\title{Ordered}
\usage{
as_ordered(x)

\method{as_ordered}{default}(x)

\method{as_ordered}{factor}(x)

\method{as_ordered}{ordered}(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
An \code{ordered} vector
}
\description{
As ordered
}
\details{
Simple implementation of \code{ordered}.  If \code{x} is \code{ordered} it is simply
returned.  If \code{x} is a \code{factor} the \code{ordered} class is added.  Otherwise,
\code{x} is made into a \code{factor} with \code{\link[=fact]{fact()}} and then the \code{ordered}
class is added.
}
