% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{assign_label}
\alias{assign_label}
\title{Dataframe labels}
\usage{
assign_label(.data, ...)
}
\arguments{
\item{.data}{A dataframe}

\item{...}{One or more unquoted expressed separated by commas}
}
\description{
Assign labels to data frames
}
\examples{
## Best when run with RStudio

labs <- assign_label(iris,
                     Sepal.Length = "cms",
                     Sepal.Width  = "cms",
                     Petal.Length = "cms",
                     Petal.Width  = "cms",
                     Species      = "Iris ...")
# View(labs)
}
