% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-utils.R
\name{attached_packages}
\alias{attached_packages}
\alias{.SendAttachedPackagesToREnviron}
\alias{.RemoveAttachedPackages}
\title{Manage attached packages}
\usage{
.SendAttachedPackagesToREnviron(remove_renviron = FALSE)

.RemoveAttachedPackages(attached = NULL)
}
\arguments{
\item{remove_renviron}{Logical, if TRUE will force an update to the
.Renvironment file}

\item{attached}{An optional vector of package names}
}
\description{
Clears out attached packages and properly loads
}
\concept{startup_funs}
