% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-conversion.R
\name{base_n}
\alias{base_n}
\title{Base N conversion}
\usage{
base_n(x, from = 10, to = 10)
}
\arguments{
\item{x}{A vector of integers}

\item{from, to}{An integer base to convert to and from; \code{from} must be an
integer from \code{1} to \code{10} and \code{to} can currently only be \code{10}.}
}
\value{
The A vector of integers converted from base \code{from} to base \code{to}
}
\description{
Convert between base numbers
}
\examples{
base_n(c(24, 22, 16), from = 7)
}
