% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_more.R
\name{between_more}
\alias{between_more}
\title{Between more}
\usage{
between_more(x, left, right, type = "gele")
}
\arguments{
\item{x}{A numeric vector of values.}

\item{type}{Abrreviated type (g: greater than, ge: greater than or equal to, l: less than, le: less than or equal to)}
}
\description{
Additional functionality and expansion of `dplyr::between`
}
\examples{
between_more(10, 2, 10, "gl")
between_more(10, 2, 10, "gle")
}
\seealso{
\link[dplyr][case_when]
}
