% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between-more.R
\name{between_more}
\alias{between_more}
\title{Between more}
\usage{
between_more(x, left, right, type = c("gele", "gel", "gle", "gl"))
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values}

\item{type}{Abbreviation for the evaluation of \code{left} on \code{right} (see details)}
}
\value{
A logical vector
}
\description{
Additional functionality and expansion of \code{dplyr::between}
}
\details{
Type can be one of the below:

\describe{
\item{g}{is greater than (>)}
\item{ge}{greater than or equal to (>=)}
\item{l}{less than (<)}
\item{ls}{less than or equal to (<=)}
}
}
\examples{
between_more(10, 2, 10, "gl")
between_more(10, 2, 10, "gle")
}
\seealso{
\code{dplyr::case_when()}
}
