% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char2fact.R
\name{char2fact}
\alias{char2fact}
\alias{char2fact.default}
\alias{char2fact.character}
\alias{char2fact.factor}
\alias{char2fact.data.frame}
\title{Character to factor}
\usage{
char2fact(x, n = 5)

\method{char2fact}{default}(x, n = 5)

\method{char2fact}{character}(x, n = 5)

\method{char2fact}{factor}(x, n = 5)

\method{char2fact}{data.frame}(x, n = 5)
}
\arguments{
\item{x}{A vector of characters}

\item{n}{The limit to the number of unique values for the factor}
}
\description{
Converts characters to factors
}
\seealso{
\code{\link[=fact2char]{fact2char()}}

Other factors:
\code{\link[=fact2char]{fact2char()}}
}
\concept{factors}
