% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\alias{write_clipboard}
\alias{read_clipboard}
\title{Write to and read from the clipboard}
\usage{
write_clipboard(x = .Last.value, ...)

read_clipboard(method = c("default", "data.frame", "tibble"), ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments sent to methods}

\item{method}{Method switch for loading the clipboard}
}
\description{
Wrappers for working with the clipboard
}
