% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{cohensd2r}
\alias{cohensd2r}
\title{Cohen's D to r}
\usage{
cohensd2r(d, n1, n2 = n1, var = FALSE)
}
\arguments{
\item{d}{Cohen's d}

\item{n1, n2}{Ns for the groups}

\item{var}{Logical.}
}
\description{
Converts Cohen's D to an r value
}
