% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts_n}
\alias{counts_n}
\title{Count N}
\usage{
counts_n(x, name = "freq")
}
\arguments{
\item{x}{A data.frame in which the combination of all columns present will
be counted}

\item{name}{A name for the new column}
}
\description{
Employs \code{counts()} for combinations of multiple rows
}
