% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact}
\alias{fact}
\alias{fact.default}
\alias{fact.character}
\alias{fact.numeric}
\alias{fact.integer}
\alias{fact.Date}
\alias{fact.POSIXct}
\alias{fact.POSIXlt}
\alias{fact.logical}
\alias{fact.factor}
\alias{fact.fact}
\alias{as_values}
\alias{as.values}
\alias{values}
\alias{converter}
\alias{converter<-}
\alias{is_fact}
\alias{is.fact}
\title{Factor}
\usage{
fact(x)

\method{fact}{default}(x)

\method{fact}{character}(x)

\method{fact}{numeric}(x)

\method{fact}{integer}(x)

\method{fact}{Date}(x)

\method{fact}{POSIXct}(x)

\method{fact}{POSIXlt}(x)

\method{fact}{logical}(x)

\method{fact}{factor}(x)

\method{fact}{fact}(x)

as_values(x)

as.values(x)

values(x)

converter(x)

converter(x) <- value

is_fact(x)

is.fact(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector of equal length of \code{x} with class \code{fact} and \code{factor}.  If
\code{x} was \code{ordered}, that class is added in between.
}
\description{
Quickly create a factor
}
\details{
\code{\link[=fact]{fact()}} can be about 5 times quicker than \code{\link[base:factor]{base::factor()}} or
\code{\link[base:as.factor]{base::as.factor()}} as it doesn't bother sorting the levels for non-numeric
data or have other checks or features.  It simply converts a vector to a
factor with all unique values as levels with \code{NA}s included.

\code{\link[=fact.factor]{fact.factor()}} will perform several checks on a factor to include
\code{NA} levels and to check if the levels should be reordered to conform with
the other methods.  The \code{\link[=fact.fact]{fact.fact()}} method simple returns \code{x}.
}
\section{level orders}{


The order of the levels may be adjusted to these rules depending on the
class of \code{x}:
\describe{
\item{\code{character}}{The order of appearance}
\item{\code{numeric}/\code{integer}/\code{Date}/\code{POSIXt}}{By the numeric order}
\item{\code{logical}}{As \code{TRUE}, \code{FALSE}, then \code{NA} if present}
\item{\code{factor}}{Numeric if levels can be safely converted, otherwise as
they are}
}
}

\seealso{
\code{\link[=as_ordered]{as_ordered()}}

Other factors:
\code{\link[=char2fact]{char2fact()}},
\code{\link[=fact2char]{fact2char()}}
}
\concept{factors}
