% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R
\name{fact}
\alias{fact}
\alias{fact.default}
\alias{fact.character}
\alias{fact.numeric}
\alias{fact.logical}
\alias{fact.factor}
\alias{fact.pseudo_id}
\title{Factor}
\usage{
fact(x)

\method{fact}{default}(x)

\method{fact}{character}(x)

\method{fact}{numeric}(x)

\method{fact}{logical}(x)

\method{fact}{factor}(x)

\method{fact}{pseudo_id}(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A \code{factor} vector of equal length of \code{x}
}
\description{
Quickly create a factor
}
\details{
\code{fact()} can be about 5 times quicker than \code{factor()} or \code{as.factor()}
as it doesn't bother sorting the levels for non-numeric data or have
other checks or features.  It simply converts a vector to a factor with all
unique values as levels with \code{NA}s included.
}
