% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{file_info}
\alias{file_info}
\alias{newest_file}
\alias{newest_dir}
\alias{oldest_file}
\alias{oldest_dir}
\alias{largest_file}
\alias{smallest_file}
\title{File information utils}
\usage{
newest_file(x)

newest_dir(x)

oldest_file(x)

oldest_dir(x)

largest_file(x)

smallest_file(x)
}
\arguments{
\item{x}{A vector of file paths}
}
\value{
A full file path
}
\description{
Other utility functions for dealing with files
}
