% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{format_to_regex}
\alias{format_to_regex}
\title{Format string to a regular expression}
\usage{
format_to_regex(x)
}
\arguments{
\item{x}{A date format, assuming}
}
\description{
Format string to a regular expression
}
\examples{
\dontrun{
format_to_regex("\%Y-\%m-\%d")
format_to_regex("\%b/\%d/\%y")
format_to_regex("\%d \%B \%Y")

pattern <- format_to_regex("\%Y-\%m-\%d")
grepl(pattern, Sys.Date())
}
}
