% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_dir_recent_date}
\alias{get_dir_recent_date}
\title{Get recent directory by date}
\usage{
get_dir_recent_date(x = ".", dt_pattern = NULL, dt_format = NULL, all = FALSE)
}
\arguments{
\item{x}{A directory}

\item{dt_pattern}{A pattern to be passed to filter for the directory}

\item{dt_format}{One or more formats to try}

\item{all}{Logical, if \code{TRUE} will recursively search for directories}
}
\value{
A full path to a directory
}
\description{
Looks at the directories and assumes the date
}
