% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_recent_dir}
\alias{get_recent_dir}
\title{Get recent directory}
\usage{
get_recent_dir(x = ".", ...)
}
\arguments{
\item{x}{The root directory}

\item{...}{Additional arguments passed to \code{\link[=list_dirs]{list_dirs()}}}
}
\value{
The full path of the most recent directory
}
\description{
Finds the recent subdirectory in a directory.
}
