% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Get recent file}
\usage{
get_recent_file(dir, pattern = "^[~$]{0}.*\\\\.csv|\\\\.txt$")
}
\arguments{
\item{dir}{The directory in which to search the file}

\item{pattern}{The regularly expression to be passed to to the file.  THe default searches for any .txt or .csv file that isn't temporary}
}
\value{
The full name of the most recent file from the stated directory
}
\description{
A function where you can detect the most recent file from a directory.
}
