% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Get recent file}
\usage{
get_recent_file(
  dir,
  pattern = NULL,
  negate = FALSE,
  exclude_temp = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{dir}{The directory in which to search the file}

\item{pattern}{The regularly expression to be passed to to the file}

\item{negate}{Logical, if TRUE, files with matching \code{pattern} will be removed}

\item{exclude_temp}{Logical, if TRUE files that begin with "^\\~\\$" are excluded}

\item{recursive}{Logical, passed to \code{list.files(., recursive)}}
}
\value{
The full name of the most recent file from the stated directory
}
\description{
A function where you can detect the most recent file from a directory.
}
