% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{get_recent_file}
\alias{get_recent_file}
\title{Get recent file}
\usage{
get_recent_file(x, exclude_temp = TRUE, ...)
}
\arguments{
\item{x}{The directory in which to search the file}

\item{exclude_temp}{Logical, if \code{TRUE} tries to remove temp Windows files}

\item{...}{Additional arguments passed to \code{\link[=list_files]{list_files()}}}
}
\value{
The full name of the most recent file from the stated directory
}
\description{
A function where you can detect the most recent file from a directory.
}
