% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{\%out\%}
\alias{\%out\%}
\alias{\%w/o\%}
\alias{no_match}
\title{Value Non-matching}
\usage{
x \%out\% table

x \%w/o\% table

no_match(x, table)
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\description{
Non matching alternatives and supplementary functions.  Contrast with [base::match()].
}
\examples{
1:10 \%in\% c(1,3,5,9)
1:10 \%out\% c(1,3,5,9)

#' (1:10) \%w/o\% c(3,7,12)
## Note that setdiff() is very similar and typically makes more sense:
        c(1:6,7:2) \%w/o\% c(3,7,12)  # -> keeps duplicates
setdiff(c(1:6,7:2),      c(3,7,12)) # -> unique values
}
