% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksource.R
\name{ksource}
\alias{ksource}
\title{Source an Rmd file}
\usage{
ksource(x, ..., quiet = TRUE)
}
\arguments{
\item{x}{An Rmd file.}

\item{...}{Additional arguments passes to [knitr::purl()]}

\item{quiet}{Logical.  Determines whether to apply silence to `knitr::purl`}
}
\description{
Sources an Rmd file as a temp file
}
