% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{assign_label}
\alias{assign_label.default}
\alias{assign_label.data.frame}
\alias{get_labels}
\alias{get_labels.data.frame}
\alias{get_labels.default}
\alias{view_labels}
\title{Dataframe labels}
\usage{
assign_label(x, ...)

\method{assign_label}{default}(x, label, ...)

\method{assign_label}{data.frame}(x, ...)

get_labels(x)

\method{get_labels}{data.frame}(x)

\method{get_labels}{default}(x)

view_labels(x, title)
}
\arguments{
\item{x}{A vector of data.frame}

\item{...}{One or more unquoted expressed separated by commas.  If assigning
to a data.frame, `...` can be replaced with a `data.frame` where the first
column is the targeted colname and the second is the desired label.}

\item{label}{A single length string of a label to be assigned}

\item{title}{Title for the viewer window -- if not supplemented will show as
`paste0(as.character(substitute(x)), " - Labels")`}
}
\description{
Assign labels to a vector or data.frame.
}
\examples{
## Best when run with RStudio

labs <- assign_label(iris,
                     Sepal.Length = "cms",
                     Sepal.Width  = "cms",
                     Petal.Length = "cms",
                     Petal.Width  = "cms",
                     Species      = "Iris ...")
# View(labs)

labs$dummy <- ""
get_labels(labs) # shows label as <NA> for dummy column
# view_labels(labs)
}
