% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{assign_labels}
\alias{assign_labels.default}
\alias{assign_labels.data.frame}
\alias{assign_label}
\alias{get_labels}
\alias{get_labels.data.frame}
\alias{get_labels.default}
\alias{view_labels}
\alias{remove_labels}
\alias{remove_labels.default}
\alias{remove_labels.data.frame}
\title{Dataframe labels}
\usage{
assign_labels(x, ...)

\method{assign_labels}{default}(x, label, ...)

\method{assign_labels}{data.frame}(x, ...)

assign_label(x, ...)

get_labels(x)

\method{get_labels}{data.frame}(x)

\method{get_labels}{default}(x)

view_labels(x, title)

remove_labels(x, ...)

\method{remove_labels}{default}(x, ...)

\method{remove_labels}{data.frame}(x, cols, ...)
}
\arguments{
\item{x}{A vector of data.frame}

\item{...}{One or more unquoted expressed separated by commas.  If assigning
to a data.frame, \code{...} can be replaced with a \code{data.frame} where the first
column is the targeted colname and the second is the desired label.}

\item{label}{A single length string of a label to be assigned}

\item{title}{Title for the viewer window -- if not supplemented will show as
\code{paste0(as.character(substitute(x)), " - Labels")}}

\item{cols}{A character vector of column names}
}
\description{
Assign labels to a vector or data.frame.
}
\examples{
## Best when run with RStudio

labs <- assign_label(iris,
                     Sepal.Length = "cms",
                     Sepal.Width  = "cms",
                     Petal.Length = "cms",
                     Petal.Width  = "cms",
                     Species      = "Iris ...")
# View(labs)

labs$dummy <- ""
get_labels(labs) # shows label as <NA> for dummy column
# view_labels(labs)

labs0 <- remove_labels(labs, c("Sepal.Length", "Sepal.Width"))
get_labels(labs0) # No labels for Sepal.Length and Sepal.Width
}
