% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length_check}
\alias{length_check}
\alias{is_length0}
\alias{no_length}
\alias{has_length}
\title{Length checkers}
\usage{
is_length0(x)

no_length(x)

has_length(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Several length checks exist:
\itemize{
\item \code{is_length0}: Not \code{NULL} but is length \code{0}
\item \code{no_length}: Length of \code{0}
\item \code{has_length}: Length is not \code{0}
}

\emph{NB}: \code{length(NULL)} is \code{0}
}
\details{
Checks lengths
}
