% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit.R
\name{limit}
\alias{limit}
\title{Limit}
\usage{
limit(x, lower = min(x), upper = max(x))
}
\arguments{
\item{x}{A numeric vector}

\item{lower}{A lower limit (as \code{x < lower})}

\item{upper}{An upper limit (as \code{x > higher})}
}
\value{
The vector \code{x} with \code{lower} and \code{upper} as the minimum, maximum
values
}
\description{
Limit a numeric vector by lower and upper bounds
}
