% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{List to data.frame}
\usage{
list2df(.list, show_NA = F, ...)
}
\arguments{
\item{.list}{A (prefferably) named `list` with any number of values}

\item{show_NA}{Logical.  If FALSE elements without names will be listed as ""; otherwise as NA.}

\item{...}{Arguments passed to `as.data.frame()`}
}
\value{
a `data.frame` object with columns "name" and "value" for the names of the `list` and the values in each
}
\description{
Converts a list object into a a tidy data.frame
}
