% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2df.R
\name{list2df}
\alias{list2df}
\title{List to data.frame}
\usage{
list2df(x, name = "name", value = "value", show_NA = FALSE, warn = FALSE)
}
\arguments{
\item{x}{A (preferably) named `list` with any number of values}

\item{name, value}{Names of the new key and value columns, respectively}

\item{show_NA}{Logical; if FALSE elements without names will be listed as "";
otherwise as NA}

\item{warn}{Logical; if TRUE will show a warning when}
}
\value{
a `data.frame` object with columns "name" and "value" for the names
of the `list` and the values in each
}
\description{
Converts a list object into a data.frame
}
\details{
Unlike [base::list2DF()], `list2df()` tries to format the data.frame by using
  the names of the list as values rather than variables.  This creates a
  longer form list that may be more tidy.
}
\examples{
\dontrun{
x <- list(a = 1, b = 2:4, c = letters[10:20])
list2df(x, "col1", "col2", force = TRUE)
# contrast with `base::list2DF()` w
list2DF(x)
}
}
