% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{list_environments}
\alias{list_environments}
\alias{environments}
\alias{ls_all}
\alias{objects_all}
\title{List all environments and objects}
\usage{
environments()

ls_all(all.names = FALSE)

objects_all(all.names = FALSE)
}
\arguments{
\item{all.names}{a logical value.  If \code{TRUE}, all
    object names are returned.  If \code{FALSE}, names which begin with a
    \samp{.} are omitted.}
}
\description{
Functions to list out all environments and objects
}
\details{
\code{environments()} is basically a printing wrapper for \code{base::search()}

\code{ls_all()} and \code{objects_all()} print out way too much text
}
