% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark-package.R
\docType{package}
\name{mark}
\alias{mark-package}
\alias{mark}
\title{mark}
\description{
Miscellaneous, Analytic R Kernels
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=mark}
  \item \url{https://github.com/jmbarbone/mark}
  \item \url{https://jmbarbone.github.io/mark/}
  \item Report bugs at \url{https://github.com/jmbarbone/mark/issues}
}

}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
