% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-arg.R
\name{match_arg}
\alias{match_arg}
\title{Match arguments}
\usage{
match_arg(x, table)
}
\arguments{
\item{x}{An argument}

\item{table}{A table of choices}
}
\description{
This function is essentially a clear version of \code{\link[base:match.arg]{base::match.arg()}} which
produces a cleaner warning message and does not restrict the \code{table} param
to \code{character} vectors only.
}
\details{
Match arguments
}
\examples{
x <- c("apple", "banana", "orange")
match_arg("b", x)

# Produces error
tryCatch(
  match_arg("pear", x),
  error = function(e) e,
  finally = invisible()
)

foo <- function(x, op = c(1, 2, 3)) {
  op <- match_arg(op)
  x / op
}

foo(10, 3)

# Error
tryCatch(
  foo(1, 0),
  error = function(e) e,
  finally = invisible()
)
}
