% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{match_param}
\alias{match_param}
\title{Match params}
\usage{
match_param(param, choices)
}
\arguments{
\item{param}{The parameter}

\item{choices}{The available choices}
}
\description{
Much like \code{\link[base:match.arg]{base::match.arg()}} with a few key differences:
\itemize{
\item Will not perform partial matching
\item Will not return error messages with ugly quotation marks
}
}
\details{
Param matching for an argument
}
