% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-arg.R
\name{match_param}
\alias{match_param}
\title{Match params}
\usage{
match_param(param, choices, null = TRUE)
}
\arguments{
\item{param}{The parameter}

\item{choices}{The available choices}

\item{null}{If \code{TRUE} allows \code{NULL} to be passed a \code{param}}
}
\value{
A single value from \code{param} matched on \code{choices}
}
\description{
Much like \code{\link[base:match.arg]{base::match.arg()}} with a few key differences:
\itemize{
\item Will not perform partial matching
\item Will not return error messages with ugly quotation marks
}
}
\details{
Param matching for an argument
}
\seealso{
\code{\link[=match_arg]{match_arg()}}
}
