% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md5.R
\name{md5}
\alias{md5}
\alias{md5.character}
\alias{md5.default}
\title{Compute the MD5 hash of an object}
\usage{
md5(x)

\method{md5}{character}(x)

\method{md5}{default}(x)
}
\arguments{
\item{x}{An object}
}
\value{
A single \code{character}
}
\description{
Compute the MD5 hash of an object
}
\examples{
md5("hello")
md5(1:10)
md5(data.frame(a = 1:10, b = letters[1:10]))
}
