% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{na_cols}
\alias{na_cols}
\alias{select_na_cols}
\alias{remove_na_cols}
\alias{is_na_cols}
\title{Selecting NA columns}
\usage{
select_na_cols(x)

remove_na_cols(x)

is_na_cols(x, names = TRUE)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{names}{Logical, if \code{TRUE} (default) will return column names as names
of vector}
}
\value{
\itemize{
\item \code{select_na_cols()} \code{x} with only columns that are all \code{NA}
\item \code{remove_na_cols()} \code{x} without columns of only \code{NA}
\item \code{is_na_cols()} a logical vector: \code{TRUE} all rows of column are \code{NA},
otherwise \code{FALSE}
}
}
\description{
Select or remove columns that are entirely NA
}
