% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{odds2r}
\alias{odds2r}
\title{Odds ratio to r}
\usage{
odds2r(odds, n1, n2 = n1)
}
\arguments{
\item{odds}{A single odds ratio.}

\item{n1, n2}{The ns of the groups.}
}
\description{
Transforms an odds ratio to an r value
}
