% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect-sizes.R
\name{odds_ratio}
\alias{odds_ratio}
\title{Odds ratio}
\usage{
odds_ratio(a, b, c, d, type = "hits_misses")
}
\arguments{
\item{.x}{A vector of values to repeat}
}
\value{
The odds ratio
}
\description{
Returns the odds ratio without returning infinity
}
