% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-values.R
\name{p_values}
\alias{p_values}
\alias{p_round}
\alias{p_value_sig}
\title{Round p-value}
\usage{
p_round(x, n = 3, sig = n)

p_value_sig(x, cutoffs = c(`***` = 0.001, `**` = 0.01, `*` = 0.05, . = 0.1))
}
\arguments{
\item{x}{A vector of p-values}

\item{n}{Number of digits to round by (if NULL - no rounding occurs)}

\item{sig}{Number of significant figures (if NULL - not used)}

\item{cutoffs}{A named vector for significant cutoffs}
}
\description{
Rounds a p-value by decimal places and reports with sig figs
}
\examples{
set.seed(42)
x <- stats::pchisq(abs(runif(25)), 4)
print(data.frame(x = x,
                 p = p_round(x),
                 sigs = p_value_sig(x)),
      digits = 3,
      right = FALSE)
}
