% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{package_available}
\alias{package_available}
\title{Check if package is available}
\usage{
package_available(namespace)
}
\arguments{
\item{namespace}{One or more packages to to require.}
}
\value{
\itemize{
\item \code{require_namespace()}: None, called for side effects
\item \code{package_available()}: Visibly, \code{TRUE} or \code{FALSE}
}
}
\description{
A wrapped requireNamespace
}
