% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-utils.R
\name{path_norm}
\alias{path_norm}
\alias{file_path}
\alias{user_file}
\title{Normalize paths}
\usage{
path_norm(x = ".", check = FALSE, remove = check)

file_path(..., check = FALSE, remove = check)

user_file(..., check = FALSE, remove = check)
}
\arguments{
\item{x}{A vector of paths}

\item{check}{Logical, if TRUE will check if the path exists and output a
warning if it does not.}

\item{remove}{Logical, if TRUE will remove paths that are not found}

\item{...}{Character vectors for creating a path}
}
\description{
Normalize and check a vector of paths
}
\examples{
\dontrun{
data(iris)
file_path <- user_file("outputs/tests.csv")
write.csv(iris, file = file_path)
}
}
