% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\alias{proportion.default}
\alias{proportion.data.frame}
\title{Proportions}
\usage{
proportion(x, ...)

\method{proportion}{default}(x, ...)

\method{proportion}{data.frame}(x, col, ...)
}
\arguments{
\item{x}{A data.frame or a vector.}

\item{...}{Additional arguments to be passed to methods.}

\item{col}{A character string of the column name which holds the groups}
}
\description{
Calculates a proportion from a vector or data.frame/matrix.
}
\examples{
proportion(iris, "Species")
proportion(iris$Species)

}
