% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Calculate a proportion}
\usage{
proportion(x, ...)
}
\arguments{
\item{x}{A data.frame or a vector.}

\item{...}{Additional arguments to be passed to methods.}
}
\description{
Calculates a proportion!
}
