% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Calculate a proportion}
\usage{
proportion(x, .name)
}
\arguments{
\item{x}{A data.frame or a vector.}

\item{.name}{The name of the column or new name for the group.  For data.frames this must be set.}

\item{...}{Additional arguments to be passed to methods (not in use).}
}
\description{
Calculates a proportion!
}
\examples{
proportion(iris, Species)
proportion(iris$Species)
proportion(iris$Species)

}
