% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{quick_df}
\alias{quick_df}
\title{Quick DF}
\usage{
quick_df(x)
}
\arguments{
\item{x}{A list}
}
\value{
A \code{data.frame}
}
\description{
This is a speedier implementation of \code{as.data.frame()} but does not provide
the same sort of checks. It should be used with caution.
}
\examples{

# unnamed will use make.names()
x <- list(1:10, letters[1:10])
quick_df(x)

# named is preferred
names(x) <- c("numbers", "letters")
quick_df(x)

}
