% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact.R, R/reexports-fuj.R,
%   R/reexports-magrittr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.ordered}
\alias{as.factor}
\alias{\%||\%}
\alias{\%:::\%}
\alias{\%::\%}
\alias{\%colons\%}
\alias{\%len\%}
\alias{\%names\%}
\alias{\%out\%}
\alias{\%wi\%}
\alias{\%wo\%}
\alias{any_match}
\alias{exattr}
\alias{flip}
\alias{is_linux}
\alias{is_macos}
\alias{is_windows}
\alias{set_names}
\alias{is_named}
\alias{remove_names}
\alias{muffle}
\alias{wuffle}
\alias{no_match}
\alias{quick_df}
\alias{quick_dfl}
\alias{require_namespace}
\alias{struct}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fuj}{\code{\link[fuj:\%::\%]{\%::\%}}, \code{\link[fuj:\%:::\%]{\%:::\%}}, \code{\link[fuj:\%colons\%]{\%colons\%}}, \code{\link[fuj:\%len\%]{\%len\%}}, \code{\link[fuj:\%names\%]{\%names\%}}, \code{\link[fuj:\%out\%]{\%out\%}}, \code{\link[fuj:\%wi\%]{\%wi\%}}, \code{\link[fuj:\%wo\%]{\%wo\%}}, \code{\link[fuj:\%||\%]{\%||\%}}, \code{\link[fuj:any_match]{any_match()}}, \code{\link[fuj:exattr]{exattr()}}, \code{\link[fuj:flip]{flip()}}, \code{\link[fuj:is_linux]{is_linux()}}, \code{\link[fuj:is_macos]{is_macos()}}, \code{\link[fuj:is_named]{is_named()}}, \code{\link[fuj:is_windows]{is_windows()}}, \code{\link[fuj:muffle]{muffle()}}, \code{\link[fuj:no_match]{no_match()}}, \code{\link[fuj:quick_df]{quick_df()}}, \code{\link[fuj:quick_dfl]{quick_dfl()}}, \code{\link[fuj:remove_names]{remove_names()}}, \code{\link[fuj:require_namespace]{require_namespace()}}, \code{\link[fuj:set_names]{set_names()}}, \code{\link[fuj:struct]{struct()}}, \code{\link[fuj:wuffle]{wuffle()}}}

  \item{generics}{\code{\link[generics:as.factor]{as.factor()}}, \code{\link[generics:as.ordered]{as.ordered()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}

