% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{pseudo_id}
\alias{fact}
\alias{as_ordered}
\alias{drop_levels}
\alias{fact_na}
\alias{fact_reverse}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{facts}{\code{\link[facts]{as_ordered}}, \code{\link[facts]{drop_levels}}, \code{\link[facts]{fact}}, \code{\link[facts]{fact_na}}, \code{\link[facts]{fact_reverse}}, \code{\link[facts]{pseudo_id}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

