% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports-fuj.R, R/reexports-magrittr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%||\%}
\alias{\%:::\%}
\alias{\%::\%}
\alias{\%colons\%}
\alias{\%len\%}
\alias{\%names\%}
\alias{\%out\%}
\alias{\%wi\%}
\alias{\%wo\%}
\alias{any_match}
\alias{exattr}
\alias{flip}
\alias{is_linux}
\alias{is_macos}
\alias{is_windows}
\alias{set_names}
\alias{is_named}
\alias{remove_names}
\alias{muffle}
\alias{wuffle}
\alias{no_match}
\alias{quick_df}
\alias{quick_dfl}
\alias{require_namespace}
\alias{struct}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fuj}{\code{\link[fuj:colons]{\%::\%}}, \code{\link[fuj:colons]{\%:::\%}}, \code{\link[fuj:colons]{\%colons\%}}, \code{\link[fuj:if_null]{\%len\%}}, \code{\link[fuj:names]{\%names\%}}, \code{\link[fuj:match_ext]{\%out\%}}, \code{\link[fuj:match_ext]{\%wi\%}}, \code{\link[fuj:match_ext]{\%wo\%}}, \code{\link[fuj:if_null]{\%||\%}}, \code{\link[fuj:match_ext]{any_match}}, \code{\link[fuj]{exattr}}, \code{\link[fuj]{flip}}, \code{\link[fuj:os]{is_linux}}, \code{\link[fuj:os]{is_macos}}, \code{\link[fuj:names]{is_named}}, \code{\link[fuj:os]{is_windows}}, \code{\link[fuj]{muffle}}, \code{\link[fuj:match_ext]{no_match}}, \code{\link[fuj]{quick_df}}, \code{\link[fuj:quick_df]{quick_dfl}}, \code{\link[fuj:names]{remove_names}}, \code{\link[fuj]{require_namespace}}, \code{\link[fuj:names]{set_names}}, \code{\link[fuj]{struct}}, \code{\link[fuj:muffle]{wuffle}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

