% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-utils.R
\name{Reload}
\alias{Reload}
\alias{.Reload}
\alias{.Restart}
\title{Reload}
\usage{
.Reload(remove_objects = TRUE, loud = FALSE)

.Restart()
}
\description{
Reloads your session
}
\details{

}
\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{ht}()},
\code{\link{startup_funs}}
}
\concept{startup_utils}
