% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{require_namespace}
\alias{require_namespace}
\alias{package_available}
\title{Require namespace}
\usage{
require_namespace(namespace, ...)

package_available(namespace)
}
\arguments{
\item{namespace, ...}{One or more packages to to require.}
}
\value{
\itemize{
\item \code{require_namespace()}: None, called for side effects
\item \code{package_available()}: Visibly, \code{TRUE} or \code{FALSE}
}
}
\description{
A wrapped requireNamespace
}
\examples{
foo <- function() {
  require_namespace("bad_package")
  1
}

try(require_namespace("bad_package"))
try(foo())
}
