% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-bind.R
\name{row_bind}
\alias{row_bind}
\title{Row bind}
\usage{
row_bind(...)
}
\arguments{
\item{...}{A list of \code{data.frames} to be attached to each other by row}
}
\value{
A \code{data.frame} combining all the rows from \code{data.frame}s in \code{...}
and all the columns, as they appear. An empty \code{data.frame} with \code{0} columns
and \code{0} rows is returned if \code{...} has no length
}
\description{
Bind a list of \code{data.frames}
}
\seealso{
\code{\link[dplyr:bind_rows]{dplyr::bind_rows()}} \code{\link[base:cbind]{base::rbind()}}
}
