% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{set_names0}
\alias{set_names0}
\alias{remove_names}
\title{Set names}
\usage{
set_names0(x, nm = x)

remove_names(x)
}
\arguments{
\item{x}{A vector of values}

\item{nm}{A vector of names}
}
\description{
Sets or removes names
}
