% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-report.R
\name{simpleTimeReport}
\alias{simpleTimeReport}
\title{Time reports}
\usage{
simpleTimeReport(title = NULL, expr, envir = parent.frame())
}
\arguments{
\item{title}{The title to be printed}

\item{expr}{The expression to run}

\item{envir}{The environment from which to evaluate the \code{expr}}
}
\value{
A \code{reported_results}/\code{list} object containing results, outputs,
messages, warnings, and errors
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function can be used to evaluate an expression line-by-line to capture
outputs, errors, messages, and evaluation time.
}
\details{
Evaluate code and report on the time difference
}
\examples{
simpleTimeReport("example", {
  print("1")
  Sys.sleep(1)
  warning("this is a warning")
  for (i in 1:5) {
    Sys.sleep(0.5)
  }
  sample(1e6, 1e6, TRUE)
})
}
