% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup-utils.R
\docType{data}
\name{startup_funs}
\alias{startup_funs}
\alias{.LoadFunctionsFromJordan}
\alias{.RunDefaultFunctionsFromJordan}
\alias{.git_branch_prompt}
\alias{.load_pipe}
\title{Start up functions}
\format{
An object of class \code{character} of length 11.
}
\usage{
.LoadFunctionsFromJordan(env = parent.frame())

.RunDefaultFunctionsFromJordan(env = parent.frame())

startup_funs

.git_branch_prompt()

.load_pipe(env = parent.frame())
}
\arguments{
\item{env}{The environment to save objects}
}
\description{
Functions that try to assist with start ups

A collection of startup functions that require no parameters
}
\details{
These functions can be used to create an enviornment for startup.
}
\seealso{
Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{ht}()}

Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{ht}()}

Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{ht}()}

Other startup_utils: 
\code{\link{.NiceMessage}()},
\code{\link{Reload}},
\code{\link{ht}()}
}
\concept{startup_utils}
\keyword{datasets}
