% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-df.R
\name{t_df}
\alias{t_df}
\title{Data frame transpose}
\usage{
t_df(x, id = 1L, name = ".id")
}
\arguments{
\item{x}{A data.frame}

\item{id}{The identification column number}

\item{name}{The name of the identification column}
}
\description{
Transposes a data.frame
}
