% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t-df.R
\name{t_df}
\alias{t_df}
\title{Data frame transpose}
\usage{
t_df(.data.frame, .id = 1)
}
\arguments{
\item{.data.frame}{A data.frame}

\item{.id}{The identification column number.}
}
\description{
Transposes a data.frame
}
