% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean.R
\name{to_boolean}
\alias{to_boolean}
\alias{to_boolean.logical}
\alias{to_boolean.numeric}
\alias{to_boolean.character}
\alias{to_boolean.factor}
\title{To Boolean}
\usage{
to_boolean(x, ...)

\method{to_boolean}{logical}(x, ...)

\method{to_boolean}{numeric}(x, true = 1L, false = 0L, ...)

\method{to_boolean}{character}(x, true = NULL, false = NULL, ...)

\method{to_boolean}{factor}(x, true = NULL, false = NULL, ...)
}
\arguments{
\item{x}{A vector of values}

\item{...}{Additional arguments passed to methods}

\item{true}{A vector of values to convert to \code{TRUE}}

\item{false}{A vector of values to convert to \code{FALSE}}
}
\value{
A \code{logical} vector of equal length as \code{x}
}
\description{
Convert a vector to boolean/logical
}
