% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-row-names.R
\name{to_row_names}
\alias{to_row_names}
\title{To row names}
\usage{
to_row_names(.df, .row_names = 1)
}
\arguments{
\item{.df}{A data.frame}

\item{.row_names}{The numeric position of the column.}
}
\description{
Converts a column to row names
}
