% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{todos}
\alias{todos}
\title{Get TODOs}
\usage{
todos()
}
\value{
`NULL` if none are found, otherwise a data.frame with the line
  number, file name, and TODO comment.
}
\description{
Search for # TODO tags
}
\details{
Calls `git grep -in "[#] TODO"` to find any lines with a comment.
Removes any finds in the NAMESPACE
}
