% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{vap}
\alias{vap}
\alias{vap_int}
\alias{vap_dbl}
\alias{vap_chr}
\alias{vap_lgl}
\alias{vap_cplx}
\alias{vap_date}
\title{Vaps!}
\usage{
vap_int(.x, .f, ..., .nm = FALSE)

vap_dbl(.x, .f, ..., .nm = FALSE)

vap_chr(.x, .f, ..., .nm = FALSE)

vap_lgl(.x, .f, ..., .nm = FALSE)

vap_cplx(.x, .f, ..., .nm = FALSE)

vap_date(.x, .f, ..., .nm = FALSE)
}
\arguments{
\item{.x}{A vector of values}

\item{.f}{A function to apply to each element in vector \code{.x}}

\item{...}{Additional arguments passed to \code{.f}}

\item{.nm}{Logical, if \code{TRUE} returns names of \code{.x} (Note: If \code{.x} does not
have any names, they will be set to the values)}
}
\description{
Wrappers for vapply
}
\details{
These are simply wrappers for \code{\link[=vapply]{vapply()}} to shorten lines.
}
