% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2df.R
\name{vector2df}
\alias{vector2df}
\title{Vector to data.frame}
\usage{
vector2df(x, name = "name", value = "value", show_NA = FALSE)
}
\arguments{
\item{x}{A vector of values.}

\item{name, value}{Character strings for the name and value columns}

\item{show_NA}{Logical}
}
\description{
Transforms a vector (named) to a data.frame
}
