% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.R
\name{vector2df}
\alias{vector2df}
\title{Vector to data.frame}
\usage{
vector2df(x, name = "name", value = "value", show_NA)
}
\arguments{
\item{x}{A vector of values.}

\item{name, value}{Character strings for the name and value columns}

\item{show_NA}{Ignored; will trigger a warning if set}
}
\value{
A \code{data.frame} with \code{name} (optional) and \code{value} columns
}
\description{
Transforms a vector (named) to a data.frame
}
